Add Chinese, Hebrew, and Greek text to speech.  Refactor gui code.
