+++ UniqueBible.code-workspace
@@ -1,0 +1,20 @@
+{
+    "folders": [
+        {
+            "path": "."
+        }
+    ],
+    "settings": {
+        "files.exclude": {
+            "**/__pycache__": true,
+            "**/installation": true,
+            "**/marvelData": true,
+            "**/music": true,
+            "**/screenshots": true,
+            "**/selectedVerses": true,
+            "**/testing": true,
+            "**/thirdParty": true,
+            "**/video": true,
+        },
+    }
+}
