+++ gui/MainWindow.py
@@ -1,12 +1,9 @@
 import os, sys, re, config, base64, webbrowser, platform, subprocess, zipfile, gdown, requests, update, myTranslation
 from datetime import datetime
 from ast import literal_eval
-from PySide2.QtCore import QUrl, Qt, QEvent, QRegExp
-from PySide2.QtGui import QIcon, QGuiApplication, QTextCursor, QFont
-from PySide2.QtPrintSupport import QPrinter, QPrintDialog
-from PySide2.QtWidgets import (QAction, QGridLayout, QVBoxLayout, QHBoxLayout, QGroupBox, QInputDialog, QLineEdit, QMainWindow, QMessageBox, QPushButton, QToolBar, QWidget, QDialog, QFileDialog, QLabel, QFrame, QTextEdit, QProgressBar, QCheckBox, QTabWidget, QComboBox, QFontDialog, QColorDialog)
-from PySide2.QtWidgets import QSplitter
-from PySide2.QtWebEngineWidgets import QWebEnginePage, QWebEngineView
+from PySide2.QtCore import QUrl, Qt, QEvent
+from PySide2.QtGui import QIcon, QGuiApplication, QFont
+from PySide2.QtWidgets import (QAction, QInputDialog, QLineEdit, QMainWindow, QMessageBox, QPushButton, QToolBar, QWidget, QFileDialog, QLabel, QFrame, QFontDialog)
 from TextCommandParser import TextCommandParser
 from BibleVerseParser import BibleVerseParser
 from BiblesSqlite import BiblesSqlite
@@ -18,47 +15,15 @@ from ToolsSqlite import BookData, IndexesSqlite
 from translations import translations
 from shutil import copyfile, rmtree
 from distutils.dir_util import copy_tree
-from themes import Themes
-
-# Optional Features
-# [Optional] Text-to-Speech feature
-try:
-    from PySide2.QtTextToSpeech import QTextToSpeech, QVoice
-    if platform.system() == "Linux" and not config.showTtsOnLinux:
-        ttsSupport = False
-    else:
-        ttsSupport = True
-except:
-    ttsSupport = False
-    print("Text-to-speech feature is not supported on this operating system.")
-# [Optional] Chinese feature - opencc
-# It converts conversion between Traditional Chinese and Simplified Chinese.
-# To enable functions working with "opencc", install python package "opencc" first, e.g. pip3 install OpenCC.
-#try:
-#    import opencc
-#    openccSupport = True
-#except:
-#    openccSupport = False
-#    print("Chinese feature 'opencc' is disabled.  To enable it, install python package 'opencc' first, by running 'pip3 install OpenCC'.")
-# [Optional] Chinese feature - pypinyin
-# It translates Chinese characters into pinyin.
-# To enable functions working with "pypinyin", install python package "pypinyin" first, e.g. pip3 install pypinyin.
-try:
-    from pypinyin import pinyin
-    pinyinSupport = True
-except:
-    pinyinSupport = False
-    print("Chinese feature 'pypinyin' is disabled.  To enable it, install python package 'pypinyin' first, by running 'pip3 install pypinyin'.")
-# [Optional] Google-translate
-try:
-    from googletrans import Translator
-    googletransSupport = True
-except:
-    googletransSupport = False
-    print("Optional feature 'googletrans' is disabled.  To enable it, install python package 'googletrans' first. Run 'pip3 install googletrans' to install.")
-# import for testing
-if config.testing:
-    import exlbl
+from gui.Downloader import Downloader
+from gui.MoreConfigOptions import MoreConfigOptions
+from gui.ImportSettings import ImportSettings
+from gui.NoteEditor import NoteEditor
+from gui.RemoteControl import RemoteControl
+from gui.MorphDialog import MorphDialog
+from gui.YouTubePopover import YouTubePopover
+from gui.CentralWidget import CentralWidget
+from gui.imports import *
 
 class MainWindow(QMainWindow):
 
@@ -165,7 +130,7 @@ class MainWindow(QMainWindow):
             # Check for missing items. Use Google Translate to translate the missing items.  Or use default English translation to fill in missing items if internet connection is not available.
             for key, value in languages.translation.items():
                 if not key in myTranslation.translation:
-                    if googletransSupport and hasattr(myTranslation, "translationLanguage"):
+                    if config.googletransSupport and hasattr(myTranslation, "translationLanguage"):
                         try:
                             languageCode = languages.codes[myTranslation.translationLanguage]
                             myTranslation.translation[key] = Translator().translate(value, dest=languageCode).text
@@ -196,7 +161,7 @@ class MainWindow(QMainWindow):
             config.thisTranslation = languages.translation
 
     def translateInterface(self):
-        if googletransSupport:
+        if config.googletransSupport:
             if not config.userLanguage:
                 self.displayMessage("{0}\n{1}".format(config.thisTranslation["message_run"], config.thisTranslation["message_setLanguage"]))
             else:
@@ -237,6 +202,7 @@ class MainWindow(QMainWindow):
         relativePath = os.path.join("htmlResources", "theText.png")
         absolutePath = os.path.abspath(relativePath)
         baseUrl = QUrl.fromLocalFile(absolutePath)
+        config.baseUrl = baseUrl
 
     def manageRemoteControl(self, open=False):
         if open or not config.remoteControl:
@@ -473,7 +439,7 @@ class MainWindow(QMainWindow):
         menu1.addSeparator()
         menu1.addAction(QAction(config.thisTranslation["menu1_update"], self, triggered=self.updateUniqueBibleApp))
         menu1.addSeparator()
-        menu1.addAction(QAction(config.thisTranslation["menu1_remoteControl"], self, triggered=self.manageRemoteControl))
+        menu1.addAction(QAction(config.thisTranslation["menu1_remoteControl"], self, shortcut="Ctrl+O", triggered=self.manageRemoteControl))
         menu1.addSeparator()
         appIcon = QIcon(os.path.join("htmlResources", "theText.png"))
         quit_action = QAction(appIcon, config.thisTranslation["menu1_exit"], self, shortcut="Ctrl+Q", triggered=qApp.quit)
@@ -1535,6 +1501,7 @@ class MainWindow(QMainWindow):
                 if item == value[-1]:
                     self.downloadHelper(self.bibleInfo[key])
                     break
+
 
     def installMarvelCommentaries(self):
         commentaries = {
@@ -3040,7 +3007,7 @@ class MainWindow(QMainWindow):
                 config.thisTranslation["menu1_setMyLanguage"], items, items.index(userLanguage), False)
         if ok and item:
             config.userLanguage = item
-            if not googletransSupport:
+            if not config.googletransSupport:
                 self.displayMessage("{0}  'googletrans'\n{1}".format(config.thisTranslation["message_missing"], config.thisTranslation["message_installFirst"]))
 
     # Set Favourite Bible Version
@@ -3081,2447 +3048,4 @@ class MainWindow(QMainWindow):
             #print(font.key())
             config.fontChinese, *_ = font.key().split(",")
             self.reloadCurrentRecord()
-
-
-class CentralWidget(QWidget):
-
-    def __init__(self, parent):
-        super().__init__()
-        self.parent = parent
-
-        #self.html = "<h1>UniqueBible.app</h1><p>This is '<b>Main View</b>'.</p>"
-
-        #self.mainView = WebEngineView(self, "main")
-        #self.mainView.setHtml(self.html, baseUrl)
-        #self.studyView = WebEngineView(self, "study")
-        #self.studyView.setHtml("This is '<b>Study View</b>'.", baseUrl)
-        
-        self.instantSplitter = QSplitter(Qt.Vertical, parent)
-      
-        if config.landscapeMode:            
-            self.parallelSplitter = QSplitter(Qt.Horizontal)
-        else:
-            self.parallelSplitter = QSplitter(Qt.Vertical)
-
-        self.mainView = TabWidget(self, "main")
-        self.parent.mainView = self.mainView
-        for i in range(config.numberOfTab):
-            tabView = WebEngineView(self, "main")
-            self.mainView.addTab(tabView, "{1}{0}".format(i+1, config.thisTranslation["tabBible"]))
-
-        self.studyView = TabWidget(self, "study")
-        self.parent.studyView = self.studyView
-        for i in range(config.numberOfTab):
-            tabView = WebEngineView(self, "study")
-            self.studyView.addTab(tabView, "{1}{0}".format(i+1, config.thisTranslation["tabStudy"]))
-
-        self.instantView = WebEngineView(self, "instant")
-        self.instantView.setHtml("<link rel='stylesheet' type='text/css' href='css/" + config.theme + ".css'><p style='font-family:{0};'><u><b>Bottom Window</b></u><br>Display instant information on this window by hovering over verse numbers, tagged words or bible reference links.</p>".format(config.font), baseUrl)
-
-        self.parallelSplitter.addWidget(self.mainView)
-        self.parallelSplitter.addWidget(self.studyView)
-        
-        self.parallelSplitter.setHandleWidth(5)
-        self.instantSplitter.setHandleWidth(5)
-
-        self.instantSplitter.addWidget(self.parallelSplitter)
-        self.instantSplitter.addWidget(self.instantView)
-
-        # Adding signals
-        self.instantSplitter.splitterMoved.connect(self.onInstantSplitterMoved)
-        self.parallelSplitter.splitterMoved.connect(self.onParallelSplitterMoved)
-
-        self.layout = QGridLayout()
-        self.layout.setContentsMargins(0, 10, 0, 3)
-        # self.layout.setHorizontalSpacing(0)
-        # self.layout.setVerticalSpacing(2)        
-        self.layout.addWidget(self.instantSplitter)                
-        
-        self.setLayout(self.layout)       
-
-    def onInstantSplitterMoved(self, pos, index):      
-        config.instantMode = -1
-        config.iModeSplitterSizes = self.instantSplitter.sizes()
-
-    def onParallelSplitterMoved(self, pos, index):
-        config.parallelMode = -1
-        config.pModeSplitterSizes = self.parallelSplitter.sizes()
-
-    def switchLandscapeMode(self):
-        if config.landscapeMode:
-            # Switch to Landscape
-            self.parallelSplitter.setOrientation(Qt.Horizontal)
-        else:
-            #Switch to Portrait
-            self.parallelSplitter.setOrientation(Qt.Vertical)
-
-    def resizeMe(self):        
-        if config.landscapeMode:
-            self.resizeAsLandscape()
-        else:
-            self.resizeAsPortrait()        
-
-    def resizeAsLandscape(self):
-        parallelRatio = {
-            0: (1, 0),
-            1: (2, 1),
-            2: (1, 1),
-            3: (1, 2),
-        }
-
-        # The total space we have.
-        w, h = [ self.parallelSplitter.width(), self.instantSplitter.height() ]
-
-        if config.parallelMode == -1:                        
-            left, right = config.pModeSplitterSizes
-            w = left + right
-        else:
-            left, right = parallelRatio[config.parallelMode]
-            
-        self.parallelSplitter.setSizes([ w * left // (left+right), w * right // (left+right)])
-       
-        if config.parallelMode:
-            self.studyView.show()
-        else:
-            self.studyView.hide()
-
-        # Then resize instantView
-        instantRatio = {
-            0: (2, 0),
-            1: (5, 2),
-            2: (2, 2),
-        }
-
-        if config.instantMode == -1:        
-            top, bottom = config.iModeSplitterSizes
-            h = top + bottom
-        else:
-            top, bottom = instantRatio[config.instantMode]
-
-        self.instantSplitter.setSizes([ h * top // (top + bottom), h * bottom // (top + bottom)])
-        
-        if config.instantMode:
-            self.instantView.show()
-        else:
-            self.instantView.hide()
-
-    def resizeAsPortrait(self):
-               
-        parallelRatio = {
-            ( 0,  0): (10, 0, 0),
-            ( 0,  1): (10, 5, 0),
-            ( 0,  2): (5, 5, 0),
-            ( 0,  3): (5, 10, 0),
-            ( 1,  0): (10, 0, 3),            
-            ( 1,  1): (20, 10, 9),
-            ( 1,  2): (5, 5, 3),
-            ( 1,  3): (10, 20, 9),
-            ( 2,  0): (10, 0, 10),
-            ( 2,  1): (10, 5, 15),
-            ( 2,  2): (10, 10, 20),
-            ( 2,  3): (5, 10, 15),
-        }      
-
-        h = self.instantSplitter.height()
-
-        if config.instantMode == -1:
-            i1, i2 = config.iModeSplitterSizes
-            h = i1 + i2
-            bottom = i2 * h // (i1 + i2)
-            remaining = h - bottom
-
-            if config.parallelMode == -1:
-                s1, s2 = config.pModeSplitterSizes
-                top = s1 * remaining // (s1 + s2)
-                middle = s2 * remaining // (s1 + s2)
-
-            else: # i.e. config.parallelMode is not custom
-                _top, _middle, _bottom = parallelRatio[(0, config.parallelMode)]
-                top = _top * remaining // (_top + _middle)
-                middle = _middle * remaining // (_top + _middle)
-
-        else: # i.e. config.instantMode is set
-
-            if config.parallelMode == -1:
-                s1, s2 = config.pModeSplitterSizes
-                _top, _middle, _bottom = parallelRatio[(config.instantMode, 0)]
-
-                bottom = h * _bottom // (_top + _middle + _bottom)
-                remaining = h - bottom
-
-                top = s1 * remaining // (s1 + s2)
-                middle = s2 * remaining // (s1 + s2)
-
-            else:
-                _top, _middle, _bottom = parallelRatio[(config.instantMode, config.parallelMode)]
-                top = h * _top // (_top + _middle + _bottom)
-                middle = h * _middle // (_top + _middle + _bottom)
-                bottom = h * _bottom // (_top + _middle + _bottom)
-
-        self.instantSplitter.setSizes([top + middle, bottom])
-        self.parallelSplitter.setSizes([top, middle])
-   
-        if config.parallelMode:
-            self.studyView.show()
-        else:
-            self.studyView.hide()
-        
-        if config.instantMode:
-            self.instantView.show()
-        else:
-            self.instantView.hide()
-
-class TabWidget(QTabWidget):
-
-    def __init__(self, parent, name):
-        super().__init__()
-        self.parent = parent
-        self.name = name
-        self.currentChanged.connect(self.tabSelected)
-        # notes:
-        # the first tab is indexed with 0.  the rest with 1,2,3,4,etc.
-        # to get the index of current tab, print(self.currentIndex())
-        # to change the current tab, self.setCurrentWidget(self.widget(4))
-
-    def translateTextIntoUserLanguage(self, text, message):
-        self.currentWidget().translateTextIntoUserLanguage(text, message)
-
-    def setHtml(self, html, baseUrl):
-        self.currentWidget().setHtml(html, baseUrl)
-
-    def load(self, path):
-        self.currentWidget().load(path)
-
-    def tabSelected(self):
-        if self.name == "main":
-            self.parent.parent.setMainPage()
-        elif self.name == "study":
-            self.parent.parent.setStudyPage()
-
-class WebEngineView(QWebEngineView):
-
-    def __init__(self, parent, name):
-        super().__init__()
-        self.parent = parent
-        self.name = name
-       
-        # add context menu (triggered by right-clicking)
-        self.setContextMenuPolicy(Qt.ActionsContextMenu)
-        self.selectionChanged.connect(self.updateContextMenu)
-        self.addMenuActions()
-
-    def displayMessage(self, message):
-        self.parent.parent.displayMessage(message)
-
-    def updateContextMenu(self):
-        text = self.getText()
-        parser = BibleVerseParser(config.parserStandarisation)
-        book = parser.bcvToVerseReference(self.getBook(), 1, 1)[:-4]
-        del parser
-        self.searchText.setText("{1} {0}".format(text, config.thisTranslation["context1_search"]))
-        self.searchTextInBook.setText("{2} {0} > {1}".format(text, book, config.thisTranslation["context1_search"]))
-        self.searchBibleTopic.setText("{1} > {0}".format(config.topic, config.thisTranslation["menu5_topics"]))
-        self.searchBibleDictionary.setText("{1} > {0}".format(config.dictionary, config.thisTranslation["context1_dict"]))
-        self.searchBibleEncyclopedia.setText("{1} > {0}".format(config.encyclopedia, config.thisTranslation["context1_encyclopedia"]))
-        self.searchThirdDictionary.setText("{1} > {0}".format(config.thirdDictionary, config.thisTranslation["menu5_3rdDict"]))
-
-    def getText(self):
-        text = {
-            "main": config.mainText,
-            "study": config.studyText,
-            "instant": config.mainText,
-        }
-        return text[self.name]
-
-    def getBook(self):
-        book = {
-            "main": config.mainB,
-            "study": config.studyB,
-            "instant": config.mainB,
-        }
-        return book[self.name]
-
-    def addMenuActions(self):
-        copyText = QAction(self)
-        copyText.setText(config.thisTranslation["context1_copy"])
-        copyText.triggered.connect(self.copySelectedText)
-        self.addAction(copyText)
-
-        separator = QAction(self)
-        separator.setSeparator(True)
-        self.addAction(separator)
-
-        # TEXT-TO-SPEECH feature
-        if ttsSupport:
-            tts = QAction(self)
-            tts.setText(config.thisTranslation["context1_speak"])
-            tts.triggered.connect(self.textToSpeech)
-            self.addAction(tts)
-
-            separator = QAction(self)
-            separator.setSeparator(True)
-            self.addAction(separator)
-
-        # Google Translate
-        if googletransSupport:
-            if config.showGoogleTranslateEnglishOptions:
-                # Translate into English
-                translateText = QAction(self)
-                translateText.setText(config.thisTranslation["context1_english"])
-                translateText.triggered.connect(self.selectedTextToEnglish)
-                self.addAction(translateText)
-            if config.showGoogleTranslateChineseOptions:
-                # Translate into Traditional Chinese
-                translateText = QAction(self)
-                translateText.setText(config.thisTranslation["context1_tChinese"])
-                translateText.triggered.connect(self.selectedTextToTraditionalChinese)
-                self.addAction(translateText)
-                # Translate into Simplified Chinese
-                translateText = QAction(self)
-                translateText.setText(config.thisTranslation["context1_sChinese"])
-                translateText.triggered.connect(self.selectedTextToSimplifiedChinese)
-                self.addAction(translateText)
-            # Translate into User-defined Language
-            if config.userLanguage:
-                userLanguage = config.userLanguage
-            else:
-                userLanguage = config.thisTranslation["context1_my"]
-            translateText = QAction(self)
-            translateText.setText("{0} {1}".format(config.thisTranslation["context1_translate"], userLanguage))
-            translateText.triggered.connect(self.checkUserLanguage)
-            self.addAction(translateText)
-
-            separator = QAction(self)
-            separator.setSeparator(True)
-            self.addAction(separator)
-
-        # CHINESE TOOL - pinyin
-        # Convert Chinese characters into pinyin
-        if pinyinSupport:
-            pinyinText = QAction(self)
-            pinyinText.setText(config.thisTranslation["context1_pinyin"])
-            pinyinText.triggered.connect(self.pinyinSelectedText)
-            self.addAction(pinyinText)
-
-            separator = QAction(self)
-            separator.setSeparator(True)
-            self.addAction(separator)
-
-        self.searchText = QAction(self)
-        self.searchText.setText(config.thisTranslation["context1_search"])
-        self.searchText.triggered.connect(self.searchSelectedText)
-        self.addAction(self.searchText)
-
-        self.searchTextInBook = QAction(self)
-        self.searchTextInBook.setText(config.thisTranslation["context1_current"])
-        self.searchTextInBook.triggered.connect(self.searchSelectedTextInBook)
-        self.addAction(self.searchTextInBook)
-
-        searchFavouriteBible = QAction(self)
-        searchFavouriteBible.setText(config.thisTranslation["context1_favourite"])
-        searchFavouriteBible.triggered.connect(self.searchSelectedFavouriteBible)
-        self.addAction(searchFavouriteBible)
-
-        separator = QAction(self)
-        separator.setSeparator(True)
-        self.addAction(separator)
-
-        searchTextOriginal = QAction(self)
-        searchTextOriginal.setText(config.thisTranslation["context1_original"])
-        searchTextOriginal.triggered.connect(self.searchSelectedTextOriginal)
-        self.addAction(searchTextOriginal)
-
-        self.searchLexicon = QAction(self)
-        self.searchLexicon.setText(config.thisTranslation["context1_originalLexicon"])
-        self.searchLexicon.triggered.connect(self.searchHebrewGreekLexicon)
-        self.addAction(self.searchLexicon)
-
-        separator = QAction(self)
-        separator.setSeparator(True)
-        self.addAction(separator)
-
-        searchFavouriteBooks = QAction(self)
-        searchFavouriteBooks.setText(config.thisTranslation["context1_favouriteBooks"])
-        searchFavouriteBooks.triggered.connect(self.searchSearchFavouriteBooks)
-        self.addAction(searchFavouriteBooks)
-
-        searchFavouriteBooks = QAction(self)
-        searchFavouriteBooks.setText(config.thisTranslation["context1_allBooks"])
-        searchFavouriteBooks.triggered.connect(self.searchAllBooks)
-        self.addAction(searchFavouriteBooks)
-
-        separator = QAction(self)
-        separator.setSeparator(True)
-        self.addAction(separator)
-
-        searchBibleCharacter = QAction(self)
-        searchBibleCharacter.setText(config.thisTranslation["menu5_characters"])
-        searchBibleCharacter.triggered.connect(self.searchCharacter)
-        self.addAction(searchBibleCharacter)
-
-        searchBibleName = QAction(self)
-        searchBibleName.setText(config.thisTranslation["menu5_names"])
-        searchBibleName.triggered.connect(self.searchName)
-        self.addAction(searchBibleName)
-
-        searchBibleLocation = QAction(self)
-        searchBibleLocation.setText(config.thisTranslation["menu5_locations"])
-        searchBibleLocation.triggered.connect(self.searchLocation)
-        self.addAction(searchBibleLocation)
-
-        self.searchBibleTopic = QAction(self)
-        self.searchBibleTopic.setText(config.thisTranslation["menu5_topics"])
-        self.searchBibleTopic.triggered.connect(self.searchTopic)
-        self.addAction(self.searchBibleTopic)
-
-        self.searchBibleEncyclopedia = QAction(self)
-        self.searchBibleEncyclopedia.setText(config.thisTranslation["context1_encyclopedia"])
-        self.searchBibleEncyclopedia.triggered.connect(self.searchEncyclopedia)
-        self.addAction(self.searchBibleEncyclopedia)
-
-        self.searchBibleDictionary = QAction(self)
-        self.searchBibleDictionary.setText(config.thisTranslation["context1_dict"])
-        self.searchBibleDictionary.triggered.connect(self.searchDictionary)
-        self.addAction(self.searchBibleDictionary)
-
-        self.searchThirdDictionary = QAction(self)
-        self.searchThirdDictionary.setText(config.thisTranslation["menu5_3rdDict"])
-        self.searchThirdDictionary.triggered.connect(self.searchThirdPartyDictionary)
-        self.addAction(self.searchThirdDictionary)
-
-        separator = QAction(self)
-        separator.setSeparator(True)
-        self.addAction(separator)
-
-        searchBibleReferences = QAction(self)
-        searchBibleReferences.setText(config.thisTranslation["context1_extract"])
-        searchBibleReferences.triggered.connect(self.extractAllReferences)
-        self.addAction(searchBibleReferences)
-
-        copyReferences = QAction(self)
-        copyReferences.setText(config.thisTranslation["context1_copyReferences"])
-        copyReferences.triggered.connect(self.copyAllReferences)
-        self.addAction(copyReferences)
-
-        separator = QAction(self)
-        separator.setSeparator(True)
-        self.addAction(separator)
-
-        runAsCommandLine = QAction(self)
-        runAsCommandLine.setText(config.thisTranslation["context1_command"])
-        runAsCommandLine.triggered.connect(self.runAsCommand)
-        self.addAction(runAsCommandLine)
-
-    def messageNoSelection(self):
-        self.displayMessage("{0}\n{1}".format(config.thisTranslation["message_run"], config.thisTranslation["message_selectWord"]))
-
-    def messageNoTtsEngine(self):
-        self.displayMessage(config.thisTranslation["message_noSupport"])
-
-    def messageNoTtsVoice(self):
-        self.displayMessage(config.thisTranslation["message_noSupport"])
-
-    def copySelectedText(self):
-        if not self.selectedText():
-            self.messageNoSelection()
-        else:
-            self.page().triggerAction(self.page().Copy)
-
-    # Translate selected words into English
-    def selectedTextToEnglish(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            self.translateTextIntoUserLanguage(selectedText)
-
-    # Translate selected words into Traditional Chinese
-    def selectedTextToTraditionalChinese(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            self.translateTextIntoUserLanguage(selectedText, "zh-TW")
-
-    # Translate selected words into Simplified Chinese
-    def selectedTextToSimplifiedChinese(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            self.translateTextIntoUserLanguage(selectedText, "zh-CN")
-
-    # Check if config.userLanguage is set
-    def checkUserLanguage(self):
-        if config.userLanguage:
-            selectedText = self.selectedText()
-            if not selectedText:
-                self.messageNoSelection()
-            else:
-                userLanguage = Languages().codes[config.userLanguage]
-                self.translateTextIntoUserLanguage(selectedText, userLanguage)
-        else:
-            self.parent.parent.openMyLanguageDialog()
-
-    # Translate selected words into user-defined language
-    def translateTextIntoUserLanguage(self, text, userLanguage="en"):
-        try:
-            translatedText = Translator().translate(text, dest=userLanguage).text
-            if config.autoCopyGoogleTranslateOutput:
-                qApp.clipboard().setText(translatedText)
-        except:
-            translatedText = "ATTENTION! Google Translate is not accessible.  Internet connection is required for this feature."
-        self.displayMessage(translatedText)
-
-    # Translate Chinese characters into pinyin
-    def pinyinSelectedText(self):
-        if not self.selectedText():
-            self.messageNoSelection()
-        else:
-            pinyinList = pinyin(self.selectedText())
-            pinyinList = [" ".join(list) for list in pinyinList]
-            pinyinText = " ".join(pinyinList)
-            if config.autoCopyChinesePinyinOutput:
-                qApp.clipboard().setText(pinyinText)
-            self.displayMessage(pinyinText)
-
-    # TEXT-TO-SPEECH feature
-    def textToSpeech(self):
-        if ttsSupport:
-            if not self.selectedText():
-                self.messageNoSelection()
-            else:
-                engineNames = QTextToSpeech.availableEngines()
-                if engineNames:
-                    self.engine = QTextToSpeech(engineNames[0])
-                    engineVoices = self.engine.availableVoices()
-                    if engineVoices:
-                        self.engine.setVoice(engineVoices[0])
-                        self.engine.setVolume(1.0)
-                        self.engine.say(self.selectedText())
-                    else:
-                        self.messageNoTtsVoice()
-                else:
-                    self.messageNoTtsEngine()
-
-    def searchSelectedText(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            searchCommand = "SEARCH:::{0}:::{1}".format(self.getText(), selectedText)
-            self.parent.parent.textCommandChanged(searchCommand, self.name)
-
-    def searchSelectedTextInBook(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            searchCommand = "ADVANCEDSEARCH:::{0}:::Book = {1} AND Scripture LIKE '%{2}%'".format(self.getText(), self.getBook(), selectedText)
-            self.parent.parent.textCommandChanged(searchCommand, self.name)
-
-    def searchSelectedFavouriteBible(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            searchCommand = "SEARCH:::{0}:::{1}".format(config.favouriteBible, selectedText)
-            self.parent.parent.textCommandChanged(searchCommand, self.name)
-
-    def searchSelectedTextOriginal(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            searchCommand = "SEARCH:::{0}:::{1}".format("MOB", selectedText)
-            self.parent.parent.textCommandChanged(searchCommand, self.name)
-
-    def searchHebrewGreekLexicon(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            searchCommand = "LEXICON:::{0}".format(selectedText)
-            self.parent.parent.textCommandChanged(searchCommand, self.name)
-
-    def searchSearchFavouriteBooks(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            searchCommand = "SEARCHBOOK:::FAV:::{0}".format(selectedText)
-            self.parent.parent.textCommandChanged(searchCommand, self.name)
-
-    def searchAllBooks(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            searchCommand = "SEARCHBOOK:::ALL:::{0}".format(selectedText)
-            self.parent.parent.textCommandChanged(searchCommand, self.name)
-
-    def searchResource(self, module):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            searchCommand = "SEARCHTOOL:::{0}:::{1}".format(module, selectedText)
-            self.parent.parent.textCommandChanged(searchCommand, self.name)
-
-    def searchCharacter(self):
-        self.searchResource("EXLBP")
-
-    def searchName(self):
-        self.searchResource("HBN")
-
-    def searchLocation(self):
-        self.searchResource("EXLBL")
-
-    def searchTopic(self):
-        self.searchResource(config.topic)
-
-    def searchDictionary(self):
-        self.searchResource(config.dictionary)
-
-    def searchEncyclopedia(self):
-        self.searchResource(config.encyclopedia)
-
-    def searchThirdPartyDictionary(self):
-        selectedText = self.selectedText()
-        if not selectedText:
-            self.messageNoSelection()
-        else:
-            searchCommand = "SEARCHTHIRDDICTIONARY:::{0}:::{1}".format(config.thirdDictionary, selectedText)
-            self.parent.parent.textCommandChanged(searchCommand, self.name)
-
-    def extractAllReferences(self):
-        selectedText = self.selectedText()
-        parser = BibleVerseParser(config.parserStandarisation)
-        verseList = parser.extractAllReferences(selectedText, False)
-        del parser
-        if not verseList:
-            self.displayMessage(config.thisTranslation["message_noReference"])
-        else:
-            biblesSqlite = BiblesSqlite()
-            verses = biblesSqlite.readMultipleVerses(self.getText(), verseList)
-            del biblesSqlite
-            self.openPopover(html=verses)
-
-    def copyAllReferences(self):
-        selectedText = self.selectedText()
-        parser = BibleVerseParser(config.parserStandarisation)
-        verseList = parser.extractAllReferences(selectedText, False)
-        if not verseList:
-            self.displayMessage(config.thisTranslation["message_noReference"])
-        else:
-            references = "; ".join([parser.bcvToVerseReference(*verse) for verse in verseList])
-            qApp.clipboard().setText(references)
-        del parser
-
-    def runAsCommand(self):
-        selectedText = self.selectedText()
-        self.parent.parent.textCommandChanged(selectedText, "main")
-
-    def createWindow(self, windowType):
-        if windowType == QWebEnginePage.WebBrowserWindow or windowType == QWebEnginePage.WebBrowserTab:
-            self.openPopover()
-        return super().createWindow(windowType)
-
-    def openPopover(self, name="popover", html="UniqueBible.app"):
-        # image options
-        if config.exportEmbeddedImages:
-            html = self.parent.parent.exportAllImages(html)
-        if config.clickToOpenImage:
-            html = self.parent.parent.addOpenImageAction(html)
-        # format html content
-        if self.name == "main":
-            activeBCVsettings = "<script>var activeText = '{0}'; var activeB = {1}; var activeC = {2}; var activeV = {3};</script>".format(config.mainText, config.mainB, config.mainC, config.mainV)
-        elif self.name == "study":
-            activeBCVsettings = "<script>var activeText = '{0}'; var activeB = {1}; var activeC = {2}; var activeV = {3};</script>".format(config.studyText, config.studyB, config.studyC, config.studyV)
-        html = "<!DOCTYPE html><html><head><title>UniqueBible.app</title><style>body {1} font-size: {3}px; font-family:'{4}'; {2} zh {1} font-family:'{5}'; {2}</style><link rel='stylesheet' type='text/css' href='css/{7}.css'><script src='js/{7}.js'></script><script src='w3.js'></script>{6}<script>var versionList = []; var compareList = []; var parallelList = []; var diffList = []; var searchList = [];</script></head><body><span id='v0.0.0'></span>{0}</body></html>".format(html, "{", "}", config.fontSize, config.font, config.fontChinese, activeBCVsettings, config.theme)
-        self.popoverView = WebEngineViewPopover(self, name, self.name)
-        self.popoverView.setHtml(html, baseUrl)
-        self.popoverView.show()
-
-
-class YouTubePopover(QWebEngineView):
-
-    def __init__(self, parent):
-        super().__init__()
-        self.parent = parent
-        self.setWindowTitle(config.thisTranslation["menu11_youtube"])
-        self.load(QUrl("https://www.youtube.com/"))
-        self.urlString = ""
-        self.urlChanged.connect(self.videoLinkChanged)
-        # add context menu (triggered by right-clicking)
-        self.setContextMenuPolicy(Qt.ActionsContextMenu)
-        self.addMenuActions()
-
-    def videoLinkChanged(self, url):
-        self.urlString = url.toString()
-
-    def addMenuActions(self):
-        goBack = QAction(self)
-        goBack.setText(config.thisTranslation["youtube_back"])
-        goBack.triggered.connect(self.back)
-        self.addAction(goBack)
-
-        goForward = QAction(self)
-        goForward.setText(config.thisTranslation["youtube_forward"])
-        goForward.triggered.connect(self.forward)
-        self.addAction(goForward)
-
-        separator = QAction(self)
-        separator.setSeparator(True)
-        self.addAction(separator)
-
-        copyLink = QAction(self)
-        copyLink.setText(config.thisTranslation["youtube_copy"])
-        copyLink.triggered.connect(self.copy)
-        self.addAction(copyLink)
-
-        separator = QAction(self)
-        separator.setSeparator(True)
-        self.addAction(separator)
-
-        mp3 = QAction(self)
-        mp3.setText(config.thisTranslation["youtube_mp3"])
-        mp3.triggered.connect(self.downloadMp3)
-        self.addAction(mp3)
-
-        video = QAction(self)
-        video.setText(config.thisTranslation["youtube_mp4"])
-        video.triggered.connect(self.downloadVideo)
-        self.addAction(video)
-
-    def back(self):
-        self.page().triggerAction(QWebEnginePage.Back)
-
-    def forward(self):
-        self.page().triggerAction(QWebEnginePage.Forward)
-
-    def copy(self):
-        if self.urlString:
-            qApp.clipboard().setText(self.urlString)
-
-    def downloadMp3(self):
-        if not self.urlString or "/results?search_query=" in self.urlString:
-            self.parent.displayMessage(config.thisTranslation["message_invalid"])
-        else:
-            self.parent.runTextCommand("mp3:::{0}".format(self.urlString))
-
-    def downloadVideo(self):
-        if not self.urlString or "/results?search_query=" in self.urlString:
-            self.parent.displayMessage(config.thisTranslation["message_invalid"])
-        else:
-            self.parent.runTextCommand("mp4:::{0}".format(self.urlString))
-
-
-class WebEngineViewPopover(QWebEngineView):
-
-    def __init__(self, parent, name, source):
-        super().__init__()
-        self.parent = parent
-        self.name = name
-        self.source = source
-        self.setWindowTitle("Unique Bible App")
-        self.titleChanged.connect(self.popoverTextCommandChanged)
-        self.page().loadFinished.connect(self.finishViewLoading)
-
-        # add context menu (triggered by right-clicking)
-        self.setContextMenuPolicy(Qt.ActionsContextMenu)
-        self.addMenuActions()
-
-    def finishViewLoading(self):
-        # scroll to the study verse
-        self.page().runJavaScript("var activeVerse = document.getElementById('v"+str(config.studyB)+"."+str(config.studyC)+"."+str(config.studyV)+"'); if (typeof(activeVerse) != 'undefined' && activeVerse != null) { activeVerse.scrollIntoView(); activeVerse.style.color = 'red'; } else { document.getElementById('v0.0.0').scrollIntoView(); }")
-
-    def popoverTextCommandChanged(self, newTextCommand):
-        # reset document.title
-        changeTitle = "document.title = 'UniqueBible.app';"
-        self.page().runJavaScript(changeTitle)
-        # run textCommandChanged from parent
-        self.parent.parent.parent.textCommandChanged(newTextCommand, self.source)
-
-    def addMenuActions(self):
-        copyText = QAction(self)
-        copyText.setText(config.thisTranslation["context1_copy"])
-        copyText.triggered.connect(self.copySelectedText)
-        self.addAction(copyText)
-
-        runAsCommandLine = QAction(self)
-        runAsCommandLine.setText(config.thisTranslation["context1_command"])
-        runAsCommandLine.triggered.connect(self.runAsCommand)
-        self.addAction(runAsCommandLine)
-
-    def messageNoSelection(self):
-        self.parent.displayMessage("{0}\n{1}".format(config.thisTranslation["message_run"], config.thisTranslation["message_selectWord"]))
-
-    def copySelectedText(self):
-        if not self.selectedText():
-            self.messageNoSelection()
-        else:
-            self.page().triggerAction(self.page().Copy)
-
-    def runAsCommand(self):
-        selectedText = self.selectedText()
-        self.parent.parent.parent.textCommandChanged(selectedText, "main")
-
-
-class MorphDialog(QDialog):
-
-    def __init__(self, parent, items):
-        super().__init__()
-        self.parent = parent
-        lexeme, self.lexicalEntry, morphologyString = items
-
-        # morphology items
-        morphology = QWidget()
-        morphologyLayout = QVBoxLayout()
-        morphologyLayout.setContentsMargins(0, 0, 0, 0)
-
-        lexemeLabel = QLabel(lexeme)
-        morphologyLayout.addWidget(lexemeLabel)
-
-        self.checkBoxes = []
-        self.morphologyList = morphologyString.split(",")
-        for counter, value in enumerate(self.morphologyList[:-1]):
-            self.checkBoxes.append(QCheckBox(value))
-            morphologyLayout.addWidget(self.checkBoxes[counter])
-        morphology.setLayout(morphologyLayout)
-
-        # Two buttons
-        buttons = QWidget()
-        buttonsLayout = QHBoxLayout()
-        self.searchButton = QPushButton("&{0}".format(config.thisTranslation["menu5_search"]))
-        self.searchButton.clicked.connect(self.searchMorphology)
-        buttonsLayout.addWidget(self.searchButton)
-        self.cancelButton = QPushButton("&{0}".format(config.thisTranslation["message_cancel"]))
-        self.cancelButton.clicked.connect(self.close)
-        buttonsLayout.addWidget(self.cancelButton)
-        buttons.setLayout(buttonsLayout)
-
-        # set main layout & title
-        mainLayout = QVBoxLayout()
-        mainLayout.addWidget(morphology)
-        mainLayout.addWidget(buttons)
-
-        self.setLayout(mainLayout)
-        self.setWindowTitle(config.thisTranslation["message_searchMorphology"])
-
-    def searchMorphology(self):
-        command = "MORPHOLOGY:::LexicalEntry LIKE '%{0},%'".format(self.lexicalEntry.split(",")[0])
-        selectedItems = [self.morphologyList[counter] for counter, value in enumerate(self.checkBoxes) if value.isChecked()]
-        if selectedItems:
-            joinedItems = ",%".join(selectedItems)
-            command += " AND Morphology LIKE '%{0}%'".format(joinedItems)
-        self.parent.runTextCommand(command)
-        self.close()
-
-
-class RemoteControl(QWidget):
-
-    def __init__(self, parent):
-        super().__init__()
-        self.setWindowTitle("Remote Control")
-        self.parent = parent
-        # specify window size
-        self.resizeWindow(2/3, 2/3)
-        # setup interface
-        self.setupUI()
-
-    # window appearance
-    def resizeWindow(self, widthFactor, heightFactor):
-        availableGeometry = qApp.desktop().availableGeometry()
-        self.resize(availableGeometry.width() * widthFactor, availableGeometry.height() * heightFactor)
-
-    # re-implementing close event, when users close this widget
-    # avoid closing by mistake
-    # this window can be closed via "Remote Control [On / Off]" in menu bar
-    def closeEvent(self, event):
-        event.ignore()
-
-    # setup ui
-    def setupUI(self):
-        mainLayout = QVBoxLayout()
-
-        self.searchLineEdit = QLineEdit()
-        self.searchLineEdit.setToolTip("Enter command here ...")
-        self.searchLineEdit.returnPressed.connect(self.searchLineEntered)
-        mainLayout.addWidget(self.searchLineEdit)
-
-        parser = BibleVerseParser(config.parserStandarisation)
-        self.bookMap = parser.standardAbbreviation
-        bookNums = list(self.bookMap.keys())
-        bookNumGps = [
-            bookNums[0:10],
-            bookNums[10:20],
-            bookNums[20:30],
-            bookNums[30:39],
-            bookNums[39:49],
-            bookNums[49:59],
-            bookNums[59:66],
-        ]
-        actionMap = {
-            "1": self.updateCommand1,
-            "2": self.updateCommand2,
-            "3": self.updateCommand3,
-            "4": self.updateCommand4,
-            "5": self.updateCommand5,
-            "6": self.updateCommand6,
-            "7": self.updateCommand7,
-            "8": self.updateCommand8,
-            "9": self.updateCommand9,
-            "10": self.updateCommand10,
-            "11": self.updateCommand11,
-            "12": self.updateCommand12,
-            "13": self.updateCommand13,
-            "14": self.updateCommand14,
-            "15": self.updateCommand15,
-            "16": self.updateCommand16,
-            "17": self.updateCommand17,
-            "18": self.updateCommand18,
-            "19": self.updateCommand19,
-            "20": self.updateCommand20,
-            "21": self.updateCommand21,
-            "22": self.updateCommand22,
-            "23": self.updateCommand23,
-            "24": self.updateCommand24,
-            "25": self.updateCommand25,
-            "26": self.updateCommand26,
-            "27": self.updateCommand27,
-            "28": self.updateCommand28,
-            "29": self.updateCommand29,
-            "30": self.updateCommand30,
-            "31": self.updateCommand31,
-            "32": self.updateCommand32,
-            "33": self.updateCommand33,
-            "34": self.updateCommand34,
-            "35": self.updateCommand35,
-            "36": self.updateCommand36,
-            "37": self.updateCommand37,
-            "38": self.updateCommand38,
-            "39": self.updateCommand39,
-            "40": self.updateCommand40,
-            "41": self.updateCommand41,
-            "42": self.updateCommand42,
-            "43": self.updateCommand43,
-            "44": self.updateCommand44,
-            "45": self.updateCommand45,
-            "46": self.updateCommand46,
-            "47": self.updateCommand47,
-            "48": self.updateCommand48,
-            "49": self.updateCommand49,
-            "50": self.updateCommand50,
-            "51": self.updateCommand51,
-            "52": self.updateCommand52,
-            "53": self.updateCommand53,
-            "54": self.updateCommand54,
-            "55": self.updateCommand55,
-            "56": self.updateCommand56,
-            "57": self.updateCommand57,
-            "58": self.updateCommand58,
-            "59": self.updateCommand59,
-            "60": self.updateCommand60,
-            "61": self.updateCommand61,
-            "62": self.updateCommand62,
-            "63": self.updateCommand63,
-            "64": self.updateCommand64,
-            "65": self.updateCommand65,
-            "66": self.updateCommand66,
-        }
-
-        otBooks = QGroupBox("Old Testament")
-        otLayout = QVBoxLayout()
-        for bookNumGp in bookNumGps[0:4]:
-            gp = QWidget()
-            layout = QHBoxLayout()
-            for bookNum in bookNumGp:
-                button = QPushButton(self.bookMap[bookNum])
-                button.clicked.connect(actionMap[bookNum])
-                layout.addWidget(button)
-            gp.setLayout(layout)
-            otLayout.addWidget(gp)
-        otBooks.setLayout(otLayout)
-        mainLayout.addWidget(otBooks)
-
-        ntBooks = QGroupBox("New Testament")
-        ntLayout = QVBoxLayout()
-        for bookNumGp in bookNumGps[4:7]:
-            gp = QWidget()
-            layout = QHBoxLayout()
-            for bookNum in bookNumGp:
-                button = QPushButton(self.bookMap[bookNum])
-                button.clicked.connect(actionMap[bookNum])
-                layout.addWidget(button)
-            gp.setLayout(layout)
-            ntLayout.addWidget(gp)
-        ntBooks.setLayout(ntLayout)
-        mainLayout.addWidget(ntBooks)
-
-        self.setLayout(mainLayout)
-
-    # search field entered
-    def searchLineEntered(self):
-        searchString = self.searchLineEdit.text()
-        self.parent.runTextCommand(searchString)
-
-    # modify search field
-    def updateCommand(self, book):
-        self.searchLineEdit.setText("{0} ".format(self.bookMap[book]))
-        self.searchLineEdit.setFocus()
-
-    def updateCommand1(self):
-        self.updateCommand("1")
-
-    def updateCommand2(self):
-        self.updateCommand("2")
-
-    def updateCommand3(self):
-        self.updateCommand("3")
-
-    def updateCommand4(self):
-        self.updateCommand("4")
-
-    def updateCommand5(self):
-        self.updateCommand("5")
-
-    def updateCommand6(self):
-        self.updateCommand("6")
-
-    def updateCommand7(self):
-        self.updateCommand("7")
-
-    def updateCommand8(self):
-        self.updateCommand("8")
-
-    def updateCommand9(self):
-        self.updateCommand("9")
-
-    def updateCommand10(self):
-        self.updateCommand("10")
-
-    def updateCommand11(self):
-        self.updateCommand("11")
-
-    def updateCommand12(self):
-        self.updateCommand("12")
-
-    def updateCommand13(self):
-        self.updateCommand("13")
-
-    def updateCommand14(self):
-        self.updateCommand("14")
-
-    def updateCommand15(self):
-        self.updateCommand("15")
-
-    def updateCommand16(self):
-        self.updateCommand("16")
-
-    def updateCommand17(self):
-        self.updateCommand("17")
-
-    def updateCommand18(self):
-        self.updateCommand("18")
-
-    def updateCommand19(self):
-        self.updateCommand("19")
-
-    def updateCommand20(self):
-        self.updateCommand("20")
-
-    def updateCommand21(self):
-        self.updateCommand("21")
-
-    def updateCommand22(self):
-        self.updateCommand("22")
-
-    def updateCommand23(self):
-        self.updateCommand("23")
-
-    def updateCommand24(self):
-        self.updateCommand("24")
-
-    def updateCommand25(self):
-        self.updateCommand("25")
-
-    def updateCommand26(self):
-        self.updateCommand("26")
-
-    def updateCommand27(self):
-        self.updateCommand("27")
-
-    def updateCommand28(self):
-        self.updateCommand("28")
-
-    def updateCommand29(self):
-        self.updateCommand("29")
-
-    def updateCommand30(self):
-        self.updateCommand("30")
-
-    def updateCommand31(self):
-        self.updateCommand("31")
-
-    def updateCommand32(self):
-        self.updateCommand("32")
-
-    def updateCommand33(self):
-        self.updateCommand("33")
-
-    def updateCommand34(self):
-        self.updateCommand("34")
-
-    def updateCommand35(self):
-        self.updateCommand("35")
-
-    def updateCommand36(self):
-        self.updateCommand("36")
-
-    def updateCommand37(self):
-        self.updateCommand("37")
-
-    def updateCommand38(self):
-        self.updateCommand("38")
-
-    def updateCommand39(self):
-        self.updateCommand("39")
-
-    def updateCommand40(self):
-        self.updateCommand("40")
-
-    def updateCommand41(self):
-        self.updateCommand("41")
-
-    def updateCommand42(self):
-        self.updateCommand("42")
-
-    def updateCommand43(self):
-        self.updateCommand("43")
-
-    def updateCommand44(self):
-        self.updateCommand("44")
-
-    def updateCommand45(self):
-        self.updateCommand("45")
-
-    def updateCommand46(self):
-        self.updateCommand("46")
-
-    def updateCommand47(self):
-        self.updateCommand("47")
-
-    def updateCommand48(self):
-        self.updateCommand("48")
-
-    def updateCommand49(self):
-        self.updateCommand("49")
-
-    def updateCommand50(self):
-        self.updateCommand("50")
-
-    def updateCommand51(self):
-        self.updateCommand("51")
-
-    def updateCommand52(self):
-        self.updateCommand("52")
-
-    def updateCommand53(self):
-        self.updateCommand("53")
-
-    def updateCommand54(self):
-        self.updateCommand("54")
-
-    def updateCommand55(self):
-        self.updateCommand("55")
-
-    def updateCommand56(self):
-        self.updateCommand("56")
-
-    def updateCommand57(self):
-        self.updateCommand("57")
-
-    def updateCommand58(self):
-        self.updateCommand("58")
-
-    def updateCommand59(self):
-        self.updateCommand("59")
-
-    def updateCommand60(self):
-        self.updateCommand("60")
-
-    def updateCommand61(self):
-        self.updateCommand("61")
-
-    def updateCommand62(self):
-        self.updateCommand("62")
-
-    def updateCommand63(self):
-        self.updateCommand("63")
-
-    def updateCommand64(self):
-        self.updateCommand("64")
-
-    def updateCommand65(self):
-        self.updateCommand("65")
-
-    def updateCommand66(self):
-        self.updateCommand("66")
-
-
-class NoteEditor(QMainWindow):
-
-    def __init__(self, parent, noteType, noteFileName="", b=None, c=None, v=None):
-        super().__init__()
-        self.parent, self.noteType = parent, noteType
-        self.noteFileName = noteFileName
-        if not self.noteType == "file":
-            if v:
-                self.b, self.c, self.v = b, c, v
-            else:
-                self.b, self.c, self.v = config.studyB, config.studyC, config.studyV
-
-        # default - "Rich" mode for editing
-        self.html = True
-        # default - show toolbar with formatting items
-        self.showToolBar = True
-        # default - text is not modified; no need for saving new content
-        self.parent.noteSaved = True
-        config.noteOpened = True
-        config.lastOpenedNote = (noteType, b, c, v)
-
-        # specify window size
-        self.resizeWindow(2/3, 2/3)
-
-        # setup interface
-        self.setupMenuBar()
-        self.addToolBarBreak()
-        self.setupToolBar()
-        self.setupLayout()
-
-        # display content when first launched
-        self.displayInitialContent()
-
-        # specify window title
-        self.updateWindowTitle()
-
-    # re-implementing close event, when users close this widget
-    def closeEvent(self, event):
-        if self.parent.noteSaved:
-            config.noteOpened = False
-            event.accept()
-        else:
-            if self.parent.warningNotSaved():
-                self.parent.noteSaved = True
-                config.noteOpened = False
-                event.accept()
-            else:
-                event.ignore()
-
-    # re-implement keyPressEvent, control+S for saving file
-    def keyPressEvent(self, event):
-        keys = {
-            Qt.Key_O: self.openFileDialog,
-            Qt.Key_S: self.saveNote,
-            Qt.Key_B: self.format_bold,
-            Qt.Key_I: self.format_italic,
-            Qt.Key_U: self.format_underline,
-            Qt.Key_M: self.format_custom,
-            Qt.Key_D: self.format_clear,
-            Qt.Key_F: self.focusSearchField,
-        }
-        key = event.key()
-        if event.modifiers() == Qt.ControlModifier and key in keys:
-            keys[key]()
-
-    # window appearance
-    def resizeWindow(self, widthFactor, heightFactor):
-        availableGeometry = qApp.desktop().availableGeometry()
-        self.resize(availableGeometry.width() * widthFactor, availableGeometry.height() * heightFactor)
-
-    def updateWindowTitle(self):
-        if self.noteType == "file":
-            if self.noteFileName:
-                *_, title = os.path.split(self.noteFileName)
-            else:
-                title = "NEW"
-        else:
-            title = self.parent.bcvToVerseReference(self.b, self.c, self.v)
-            if self.noteType == "chapter":
-                title, *_ = title.split(":")
-        mode = {True: "rich", False: "plain"}
-        notModified = {True: "", False: " [modified]"}
-        self.setWindowTitle("Note Editor ({1} mode) - {0}{2}".format(title, mode[self.html], notModified[self.parent.noteSaved]))
-
-    # switching between "rich" & "plain" mode
-    def switchMode(self):
-        if self.html:
-            note = self.editor.toHtml()
-            note = re.sub("<body style={0}[ ]*?font-family:[ ]*?'[^']*?';[ ]*?font-size:[ ]*?[0-9]+?pt;".format('"'), "<body style={0}font-family:'{1}'; font-size:{2}pt;".format('"', config.font, config.fontSize), note)
-            self.editor.setPlainText(note)
-            self.html = False
-            self.updateWindowTitle()
-        else:
-            note = self.editor.toPlainText()
-            self.editor.setHtml(note)
-            self.html = True
-            self.updateWindowTitle()
-        # without this hide / show command below, QTextEdit does not update the text in some devices
-        self.hide()
-        self.show()
-
-    def setupMenuBar(self):
-        if config.toolBarIconFullSize:
-            self.setupMenuBarFullIconSize()
-        else:
-            self.setupMenuBarStandardIconSize()
-
-    def setupMenuBarStandardIconSize(self):
-
-        self.menuBar = QToolBar()
-        self.menuBar.setWindowTitle(config.thisTranslation["note_title"])
-        self.menuBar.setContextMenuPolicy(Qt.PreventContextMenu)
-        # In QWidget, self.menuBar is treated as the menubar without the following line
-        # In QMainWindow, the following line adds the configured QToolBar as part of the toolbar of the main window
-        self.addToolBar(self.menuBar)
-
-        newButton = QPushButton()
-        newButton.setToolTip("{0}\n[Ctrl/Cmd + N]".format(config.thisTranslation["menu7_create"]))
-        newButtonFile = os.path.join("htmlResources", "newfile.png")
-        newButton.setIcon(QIcon(newButtonFile))
-        newButton.clicked.connect(self.newNoteFile)
-        self.menuBar.addWidget(newButton)
-
-        openButton = QPushButton()
-        openButton.setToolTip("{0}\n[Ctrl/Cmd + O]".format(config.thisTranslation["menu7_open"]))
-        openButtonFile = os.path.join("htmlResources", "open.png")
-        openButton.setIcon(QIcon(openButtonFile))
-        openButton.clicked.connect(self.openFileDialog)
-        self.menuBar.addWidget(openButton)
-
-        self.menuBar.addSeparator()
-
-        saveButton = QPushButton()
-        saveButton.setToolTip("{0}\n[Ctrl/Cmd + S]".format(config.thisTranslation["note_save"]))
-        saveButtonFile = os.path.join("htmlResources", "save.png")
-        saveButton.setIcon(QIcon(saveButtonFile))
-        saveButton.clicked.connect(self.saveNote)
-        self.menuBar.addWidget(saveButton)
-
-        saveAsButton = QPushButton()
-        saveAsButton.setToolTip(config.thisTranslation["note_saveAs"])
-        saveAsButtonFile = os.path.join("htmlResources", "saveas.png")
-        saveAsButton.setIcon(QIcon(saveAsButtonFile))
-        saveAsButton.clicked.connect(self.openSaveAsDialog)
-        self.menuBar.addWidget(saveAsButton)
-
-        self.menuBar.addSeparator()
-
-        toolBarButton = QPushButton()
-        toolBarButton.setToolTip(config.thisTranslation["note_print"])
-        toolBarButtonFile = os.path.join("htmlResources", "print.png")
-        toolBarButton.setIcon(QIcon(toolBarButtonFile))
-        toolBarButton.clicked.connect(self.printNote)
-        self.menuBar.addWidget(toolBarButton)
-
-        self.menuBar.addSeparator()
-
-        switchButton = QPushButton()
-        switchButton.setToolTip(config.thisTranslation["note_mode"])
-        switchButtonFile = os.path.join("htmlResources", "switch.png")
-        switchButton.setIcon(QIcon(switchButtonFile))
-        switchButton.clicked.connect(self.switchMode)
-        self.menuBar.addWidget(switchButton)
-
-        self.menuBar.addSeparator()
-
-#        decreaseFontSizeButton = QPushButton()
-#        decreaseFontSizeButton.setToolTip(config.thisTranslation["menu2_smaller"])
-#        decreaseFontSizeButtonFile = os.path.join("htmlResources", "fontMinus.png")
-#        decreaseFontSizeButton.setIcon(QIcon(decreaseFontSizeButtonFile))
-#        decreaseFontSizeButton.clicked.connect(self.decreaseNoteEditorFontSize)
-#        self.menuBar.addWidget(decreaseFontSizeButton)
-#
-#        increaseFontSizeButton = QPushButton()
-#        increaseFontSizeButton.setToolTip(config.thisTranslation["menu2_larger"])
-#        increaseFontSizeButtonFile = os.path.join("htmlResources", "fontPlus.png")
-#        increaseFontSizeButton.setIcon(QIcon(increaseFontSizeButtonFile))
-#        increaseFontSizeButton.clicked.connect(self.increaseNoteEditorFontSize)
-#        self.menuBar.addWidget(increaseFontSizeButton)
-
-#        self.menuBar.addSeparator()
-
-        self.searchLineEdit = QLineEdit()
-        self.searchLineEdit.setToolTip(config.thisTranslation["menu5_search"])
-        self.searchLineEdit.setMaximumWidth(400)
-        self.searchLineEdit.returnPressed.connect(self.searchLineEntered)
-        self.menuBar.addWidget(self.searchLineEdit)
-
-        self.menuBar.addSeparator()
-
-        toolBarButton = QPushButton()
-        toolBarButton.setToolTip(config.thisTranslation["note_toolbar"])
-        toolBarButtonFile = os.path.join("htmlResources", "toolbar.png")
-        toolBarButton.setIcon(QIcon(toolBarButtonFile))
-        toolBarButton.clicked.connect(self.toogleToolbar)
-        self.menuBar.addWidget(toolBarButton)
-
-        self.menuBar.addSeparator()
-
-    def setupMenuBarFullIconSize(self):
-
-        self.menuBar = QToolBar()
-        self.menuBar.setWindowTitle(config.thisTranslation["note_title"])
-        self.menuBar.setContextMenuPolicy(Qt.PreventContextMenu)
-        # In QWidget, self.menuBar is treated as the menubar without the following line
-        # In QMainWindow, the following line adds the configured QToolBar as part of the toolbar of the main window
-        self.addToolBar(self.menuBar)
-
-        iconFile = os.path.join("htmlResources", "newfile.png")
-        self.menuBar.addAction(QIcon(iconFile), "{0}\n[Ctrl/Cmd + N]".format(config.thisTranslation["menu7_create"]), self.newNoteFile)
-
-        iconFile = os.path.join("htmlResources", "open.png")
-        self.menuBar.addAction(QIcon(iconFile), "{0}\n[Ctrl/Cmd + O]".format(config.thisTranslation["menu7_open"]), self.openFileDialog)
-
-        self.menuBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "save.png")
-        self.menuBar.addAction(QIcon(iconFile), "{0}\n[Ctrl/Cmd + S]".format(config.thisTranslation["note_save"]), self.saveNote)
-
-        iconFile = os.path.join("htmlResources", "saveas.png")
-        self.menuBar.addAction(QIcon(iconFile), config.thisTranslation["note_saveAs"], self.openSaveAsDialog)
-
-        self.menuBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "print.png")
-        self.menuBar.addAction(QIcon(iconFile), config.thisTranslation["note_print"], self.printNote)
-
-        self.menuBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "switch.png")
-        self.menuBar.addAction(QIcon(iconFile), config.thisTranslation["note_mode"], self.switchMode)
-
-        self.menuBar.addSeparator()
-
-#        iconFile = os.path.join("htmlResources", "fontMinus.png")
-#        self.menuBar.addAction(QIcon(iconFile), config.thisTranslation["menu2_smaller"], self.decreaseNoteEditorFontSize)
-#
-#        iconFile = os.path.join("htmlResources", "fontPlus.png")
-#        self.menuBar.addAction(QIcon(iconFile), config.thisTranslation["menu2_larger"], self.increaseNoteEditorFontSize)
-
-#        self.menuBar.addSeparator()
-
-        self.searchLineEdit = QLineEdit()
-        self.searchLineEdit.setToolTip("{0}\n[Ctrl/Cmd + F]".format(config.thisTranslation["menu5_search"]))
-        self.searchLineEdit.setMaximumWidth(400)
-        self.searchLineEdit.returnPressed.connect(self.searchLineEntered)
-        self.menuBar.addWidget(self.searchLineEdit)
-
-        self.menuBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "toolbar.png")
-        self.menuBar.addAction(QIcon(iconFile), config.thisTranslation["note_toolbar"], self.toogleToolbar)
-
-        self.menuBar.addSeparator()
-
-    def toogleToolbar(self):
-        if self.showToolBar:
-            self.toolBar.hide()
-            self.showToolBar = False
-        else:
-            self.toolBar.show()
-            self.showToolBar = True
-
-    def printNote(self):
-        #document = QTextDocument("Sample Page")
-        document = self.editor.document()
-        printer = QPrinter()
-
-        myPrintDialog = QPrintDialog(printer, self)
-        if myPrintDialog.exec_() == QDialog.Accepted:
-            return document.print_(printer)
-
-    def setupToolBar(self):
-        if config.toolBarIconFullSize:
-            self.setupToolBarFullIconSize()
-        else:
-            self.setupToolBarStandardIconSize()
-
-    def setupToolBarStandardIconSize(self):
-
-        self.toolBar = QToolBar()
-        self.toolBar.setWindowTitle(config.thisTranslation["noteTool_title"])
-        self.toolBar.setContextMenuPolicy(Qt.PreventContextMenu)
-        # self.toolBar can be treated as an individual widget and positioned with a specified layout
-        # In QMainWindow, the following line adds the configured QToolBar as part of the toolbar of the main window
-        self.addToolBar(self.toolBar)
-
-        fontButton = QPushButton()
-        fontButton.setToolTip(config.thisTranslation["noteTool_textFont"])
-        fontButtonFile = os.path.join("htmlResources", "font.png")
-        fontButton.setIcon(QIcon(fontButtonFile))
-        fontButton.clicked.connect(self.format_font)
-        self.toolBar.addWidget(fontButton)
-
-        fontButton = QPushButton()
-        fontButton.setToolTip(config.thisTranslation["noteTool_textColor"])
-        fontButtonFile = os.path.join("htmlResources", "textColor.png")
-        fontButton.setIcon(QIcon(fontButtonFile))
-        fontButton.clicked.connect(self.format_textColor)
-        self.toolBar.addWidget(fontButton)
-
-        fontButton = QPushButton()
-        fontButton.setToolTip(config.thisTranslation["noteTool_textBackgroundColor"])
-        fontButtonFile = os.path.join("htmlResources", "textBgColor.png")
-        fontButton.setIcon(QIcon(fontButtonFile))
-        fontButton.clicked.connect(self.format_textBackgroundColor)
-        self.toolBar.addWidget(fontButton)
-
-        self.toolBar.addSeparator()
-
-        headerButton = QPushButton()
-        headerButton.setToolTip(config.thisTranslation["noteTool_header1"])
-        headerButtonFile = os.path.join("htmlResources", "header1.png")
-        headerButton.setIcon(QIcon(headerButtonFile))
-        headerButton.clicked.connect(self.format_header1)
-        self.toolBar.addWidget(headerButton)
-
-        headerButton = QPushButton()
-        headerButton.setToolTip(config.thisTranslation["noteTool_header2"])
-        headerButtonFile = os.path.join("htmlResources", "header2.png")
-        headerButton.setIcon(QIcon(headerButtonFile))
-        headerButton.clicked.connect(self.format_header2)
-        self.toolBar.addWidget(headerButton)
-
-        headerButton = QPushButton()
-        headerButton.setToolTip(config.thisTranslation["noteTool_header3"])
-        headerButtonFile = os.path.join("htmlResources", "header3.png")
-        headerButton.setIcon(QIcon(headerButtonFile))
-        headerButton.clicked.connect(self.format_header3)
-        self.toolBar.addWidget(headerButton)
-
-        self.toolBar.addSeparator()
-        
-        boldButton = QPushButton()
-        boldButton.setToolTip("{0}\n[Ctrl/Cmd + B]".format(config.thisTranslation["noteTool_bold"]))
-        boldButtonFile = os.path.join("htmlResources", "bold.png")
-        boldButton.setIcon(QIcon(boldButtonFile))
-        boldButton.clicked.connect(self.format_bold)
-        self.toolBar.addWidget(boldButton)
-
-        italicButton = QPushButton()
-        italicButton.setToolTip("{0}\n[Ctrl/Cmd + I]".format(config.thisTranslation["noteTool_italic"]))
-        italicButtonFile = os.path.join("htmlResources", "italic.png")
-        italicButton.setIcon(QIcon(italicButtonFile))
-        italicButton.clicked.connect(self.format_italic)
-        self.toolBar.addWidget(italicButton)
-
-        underlineButton = QPushButton()
-        underlineButton.setToolTip("{0}\n[Ctrl/Cmd + U]".format(config.thisTranslation["noteTool_underline"]))
-        underlineButtonFile = os.path.join("htmlResources", "underline.png")
-        underlineButton.setIcon(QIcon(underlineButtonFile))
-        underlineButton.clicked.connect(self.format_underline)
-        self.toolBar.addWidget(underlineButton)
-
-        self.toolBar.addSeparator()
-
-        customButton = QPushButton()
-        customButton.setToolTip("{0}\n[Ctrl/Cmd + M]\n\n{1}\n* {4}\n* {5}\n* {6}\n\n{2}\n*1 {4}\n*2 {5}\n*3 {6}\n\n{3}\n{10}{4}|{5}|{6}{11}\n{10}{7}|{8}|{9}{11}".format(config.thisTranslation["noteTool_trans0"], config.thisTranslation["noteTool_trans1"], config.thisTranslation["noteTool_trans2"], config.thisTranslation["noteTool_trans3"], config.thisTranslation["noteTool_no1"], config.thisTranslation["noteTool_no2"], config.thisTranslation["noteTool_no3"], config.thisTranslation["noteTool_no4"], config.thisTranslation["noteTool_no5"], config.thisTranslation["noteTool_no6"], "{", "}"))
-        customButtonFile = os.path.join("htmlResources", "custom.png")
-        customButton.setIcon(QIcon(customButtonFile))
-        customButton.clicked.connect(self.format_custom)
-        self.toolBar.addWidget(customButton)
-
-        self.toolBar.addSeparator()
-
-        leftButton = QPushButton()
-        leftButton.setToolTip(config.thisTranslation["noteTool_left"])
-        leftButtonFile = os.path.join("htmlResources", "align_left.png")
-        leftButton.setIcon(QIcon(leftButtonFile))
-        leftButton.clicked.connect(self.format_left)
-        self.toolBar.addWidget(leftButton)
-
-        centerButton = QPushButton()
-        centerButton.setToolTip(config.thisTranslation["noteTool_centre"])
-        centerButtonFile = os.path.join("htmlResources", "align_center.png")
-        centerButton.setIcon(QIcon(centerButtonFile))
-        centerButton.clicked.connect(self.format_center)
-        self.toolBar.addWidget(centerButton)
-
-        rightButton = QPushButton()
-        rightButton.setToolTip(config.thisTranslation["noteTool_right"])
-        rightButtonFile = os.path.join("htmlResources", "align_right.png")
-        rightButton.setIcon(QIcon(rightButtonFile))
-        rightButton.clicked.connect(self.format_right)
-        self.toolBar.addWidget(rightButton)
-
-        justifyButton = QPushButton()
-        justifyButton.setToolTip(config.thisTranslation["noteTool_justify"])
-        justifyButtonFile = os.path.join("htmlResources", "align_justify.png")
-        justifyButton.setIcon(QIcon(justifyButtonFile))
-        justifyButton.clicked.connect(self.format_justify)
-        self.toolBar.addWidget(justifyButton)
-
-        self.toolBar.addSeparator()
-
-        clearButton = QPushButton()
-        clearButton.setToolTip("{0}\n[Ctrl/Cmd + D]".format(config.thisTranslation["noteTool_delete"]))
-        clearButtonFile = os.path.join("htmlResources", "clearFormat.png")
-        clearButton.setIcon(QIcon(clearButtonFile))
-        clearButton.clicked.connect(self.format_clear)
-        self.toolBar.addWidget(clearButton)
-
-        self.toolBar.addSeparator()
-
-        hyperlinkButton = QPushButton()
-        hyperlinkButton.setToolTip(config.thisTranslation["noteTool_hyperlink"])
-        hyperlinkButtonFile = os.path.join("htmlResources", "hyperlink.png")
-        hyperlinkButton.setIcon(QIcon(hyperlinkButtonFile))
-        hyperlinkButton.clicked.connect(self.openHyperlinkDialog)
-        self.toolBar.addWidget(hyperlinkButton)
-
-        imageButton = QPushButton()
-        imageButton.setToolTip(config.thisTranslation["noteTool_externalImage"])
-        imageButtonFile = os.path.join("htmlResources", "gallery.png")
-        imageButton.setIcon(QIcon(imageButtonFile))
-        imageButton.clicked.connect(self.openImageDialog)
-        self.toolBar.addWidget(imageButton)
-
-        self.toolBar.addSeparator()
-
-        imageButton = QPushButton()
-        imageButton.setToolTip(config.thisTranslation["noteTool_image"])
-        imageButtonFile = os.path.join("htmlResources", "addImage.png")
-        imageButton.setIcon(QIcon(imageButtonFile))
-        imageButton.clicked.connect(self.addInternalImage)
-        self.toolBar.addWidget(imageButton)
-
-        imageButton = QPushButton()
-        imageButton.setToolTip(config.thisTranslation["noteTool_exportImage"])
-        imageButtonFile = os.path.join("htmlResources", "export.png")
-        imageButton.setIcon(QIcon(imageButtonFile))
-        imageButton.clicked.connect(self.exportNoteImages)
-        self.toolBar.addWidget(imageButton)
-
-        self.toolBar.addSeparator()
-
-    def setupToolBarFullIconSize(self):
-
-        self.toolBar = QToolBar()
-        self.toolBar.setWindowTitle(config.thisTranslation["noteTool_title"])
-        self.toolBar.setContextMenuPolicy(Qt.PreventContextMenu)
-        # self.toolBar can be treated as an individual widget and positioned with a specified layout
-        # In QMainWindow, the following line adds the configured QToolBar as part of the toolbar of the main window
-        self.addToolBar(self.toolBar)
-
-        iconFile = os.path.join("htmlResources", "font.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_textFont"], self.format_font)
-
-        iconFile = os.path.join("htmlResources", "textColor.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_textColor"], self.format_textColor)
-
-        iconFile = os.path.join("htmlResources", "textBgColor.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_textBackgroundColor"], self.format_textBackgroundColor)
-
-        self.toolBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "header1.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_header1"], self.format_header1)
-
-        iconFile = os.path.join("htmlResources", "header2.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_header2"], self.format_header2)
-
-        iconFile = os.path.join("htmlResources", "header3.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_header3"], self.format_header3)
-
-        self.toolBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "bold.png")
-        self.toolBar.addAction(QIcon(iconFile), "{0}\n[Ctrl/Cmd + B]".format(config.thisTranslation["noteTool_bold"]), self.format_bold)
-
-        iconFile = os.path.join("htmlResources", "italic.png")
-        self.toolBar.addAction(QIcon(iconFile), "{0}\n[Ctrl/Cmd + I]".format(config.thisTranslation["noteTool_italic"]), self.format_italic)
-
-        iconFile = os.path.join("htmlResources", "underline.png")
-        self.toolBar.addAction(QIcon(iconFile), "{0}\n[Ctrl/Cmd + U]".format(config.thisTranslation["noteTool_underline"]), self.format_underline)
-
-        self.toolBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "custom.png")
-        self.toolBar.addAction(QIcon(iconFile), "{0}\n[Ctrl/Cmd + M]\n\n{1}\n* {4}\n* {5}\n* {6}\n\n{2}\n*1 {4}\n*2 {5}\n*3 {6}\n\n{3}\n{10}{4}|{5}|{6}{11}\n{10}{7}|{8}|{9}{11}".format(config.thisTranslation["noteTool_trans0"], config.thisTranslation["noteTool_trans1"], config.thisTranslation["noteTool_trans2"], config.thisTranslation["noteTool_trans3"], config.thisTranslation["noteTool_no1"], config.thisTranslation["noteTool_no2"], config.thisTranslation["noteTool_no3"], config.thisTranslation["noteTool_no4"], config.thisTranslation["noteTool_no5"], config.thisTranslation["noteTool_no6"], "{", "}"), self.format_custom)
-
-        self.toolBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "align_left.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_left"], self.format_left)
-
-        iconFile = os.path.join("htmlResources", "align_center.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_centre"], self.format_center)
-
-        iconFile = os.path.join("htmlResources", "align_right.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_right"], self.format_right)
-
-        iconFile = os.path.join("htmlResources", "align_justify.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_justify"], self.format_justify)
-
-        self.toolBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "clearFormat.png")
-        self.toolBar.addAction(QIcon(iconFile), "{0}\n[Ctrl/Cmd + D]".format(config.thisTranslation["noteTool_delete"]), self.format_clear)
-
-        self.toolBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "hyperlink.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_hyperlink"], self.openHyperlinkDialog)
-
-        iconFile = os.path.join("htmlResources", "gallery.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_externalImage"], self.openImageDialog)
-
-        self.toolBar.addSeparator()
-
-        iconFile = os.path.join("htmlResources", "addImage.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_image"], self.addInternalImage)
-
-        iconFile = os.path.join("htmlResources", "export.png")
-        self.toolBar.addAction(QIcon(iconFile), config.thisTranslation["noteTool_exportImage"], self.exportNoteImages)
-
-        self.toolBar.addSeparator()
-
-    def setupLayout(self):
-        self.editor = QTextEdit()        
-        self.editor.setStyleSheet("font-family:'{0}'; font-size:{1}pt;".format(config.font, config.fontSize));
-        self.editor.textChanged.connect(self.textChanged)
-        self.setCentralWidget(self.editor)
-
-        #self.layout = QGridLayout()
-        #self.layout.setMenuBar(self.menuBar)
-        #self.layout.addWidget(self.toolBar, 0, 0)
-        #self.layout.addWidget(self.editor, 1, 0)
-        #self.setLayout(self.layout)
-
-    # adjustment of note editor font size
-    def increaseNoteEditorFontSize(self):
-        if self.html:
-            self.editor.selectAll()
-            config.noteEditorFontSize += 1
-            self.editor.setFontPointSize(config.noteEditorFontSize)
-            self.hide()
-            self.show()
-
-    def decreaseNoteEditorFontSize(self):
-        if self.html and not config.noteEditorFontSize == 0:
-            self.editor.selectAll()
-            config.noteEditorFontSize -= 1
-            self.editor.setFontPointSize(config.noteEditorFontSize)
-            self.hide()
-            self.show()
-
-    # search field entered
-    def searchLineEntered(self):
-        searchString = self.searchLineEdit.text()
-        if searchString:
-            cursor = self.editor.document().find(searchString, self.editor.textCursor())
-        if cursor:
-            self.editor.setTextCursor(cursor)
-        self.hide()
-        self.show()
-
-    def focusSearchField(self):
-        self.searchLineEdit.setFocus()
-
-    # track if the text being modified
-    def textChanged(self):
-        if self.parent.noteSaved:
-            self.parent.noteSaved = False
-            self.updateWindowTitle()
-
-    # display content when first launched
-    def displayInitialContent(self):
-        if self.noteType == "file":
-            if self.noteFileName:
-                self.openNoteFile(self.noteFileName)
-            else:
-                self.newNoteFile()
-        else:
-            self.openBibleNote()
-
-        self.editor.selectAll()
-        self.editor.setFontPointSize(config.noteEditorFontSize)
-        self.editor.moveCursor(QTextCursor.Start, QTextCursor.MoveAnchor)
-
-        self.parent.noteSaved = True
-
-    def getEmptyPage(self):
-        return """
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0//EN" "http://www.w3.org/TR/REC-html40/strict.dtd">
-<html><head><meta name="qrichtext" content="1" /><style type="text/css">
-p, li {0} white-space: pre-wrap; {1}
-</style></head><body style="font-family:'{2}'; font-size:{3}pt; font-weight:400; font-style:normal;">
-<p style="-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><br /></p></body></html>""".format("{", "}", config.font, config.fontSize)
-
-    # load chapter / verse notes from sqlite database
-    def openBibleNote(self):
-        noteSqlite = NoteSqlite()
-        if self.noteType == "chapter":
-            note = noteSqlite.getChapterNote((self.b, self.c))
-        elif self.noteType == "verse":
-            note = noteSqlite.getVerseNote((self.b, self.c, self.v))
-        del noteSqlite
-        if note == config.thisTranslation["empty"]:
-            note = self.getEmptyPage()
-        else:
-            note = self.fixNoteFont(note)
-        if self.html:
-            self.editor.setHtml(note)
-        else:
-            self.editor.setPlainText(note)
-
-    # File I / O
-    def newNoteFile(self):
-        if self.parent.noteSaved:
-            self.newNoteFileAction()
-        elif self.parent.warningNotSaved():
-            self.newNoteFileAction()
-
-    def newNoteFileAction(self):
-        self.noteType = "file"
-        self.noteFileName = ""
-        #self.editor.clear()
-        defaultText = self.getEmptyPage()
-        if self.html:
-            self.editor.setHtml(defaultText)
-        else:
-            self.editor.setPlainText(defaultText)
-        self.parent.noteSaved = True
-        self.updateWindowTitle()
-        self.hide()
-        self.show()
-
-    def openFileDialog(self):
-        if self.parent.noteSaved:
-            self.openFileDialogAction()
-        elif self.parent.warningNotSaved():
-            self.openFileDialogAction()
-
-    def openFileDialogAction(self):
-        options = QFileDialog.Options()
-        fileName, filtr = QFileDialog.getOpenFileName(self,
-                config.thisTranslation["menu7_open"],
-                self.parent.openFileNameLabel.text(),
-                "UniqueBible.app Note Files (*.uba);;HTML Files (*.html);;HTM Files (*.htm);;All Files (*)", "", options)
-        if fileName:
-            self.openNoteFile(fileName)
-
-    def openNoteFile(self, fileName):
-        try:
-            f = open(fileName, "r", encoding="utf-8")
-        except:
-            print("Failed to open '{0}'".format(fileName))
-        note = f.read()
-        f.close()
-        self.noteType = "file"
-        self.noteFileName = fileName
-        note = self.fixNoteFont(note)
-        if self.html:
-            self.editor.setHtml(note)
-        else:
-            self.editor.setPlainText(note)
-        self.parent.noteSaved = True
-        self.updateWindowTitle()
-        self.hide()
-        self.show()
-
-    def saveNote(self):
-        if self.html:
-            note = self.editor.toHtml()
-        else:
-            note = self.editor.toPlainText()
-        note = self.fixNoteFont(note)
-        if self.noteType == "chapter":
-            noteSqlite = NoteSqlite()
-            noteSqlite.saveChapterNote((self.b, self.c, note))
-            del noteSqlite
-            if config.openBibleNoteAfterSave:
-                self.parent.openChapterNote(self.b, self.c)
-            self.parent.noteSaved = True
-            self.updateWindowTitle()
-        elif self.noteType == "verse":
-            noteSqlite = NoteSqlite()
-            noteSqlite.saveVerseNote((self.b, self.c, self.v, note))
-            del noteSqlite
-            if config.openBibleNoteAfterSave:
-                self.parent.openVerseNote(self.b, self.c, self.v)
-            self.parent.noteSaved = True
-            self.updateWindowTitle()
-        elif self.noteType == "file":
-            if self.noteFileName == "":
-                self.openSaveAsDialog()
-            else:
-                self.saveAsNote(self.noteFileName)
-
-    def openSaveAsDialog(self):
-        if self.noteFileName:
-            *_, defaultName = os.path.split(self.noteFileName)
-        else:
-            defaultName = "new.uba"
-        options = QFileDialog.Options()
-        fileName, filtr = QFileDialog.getSaveFileName(self,
-                config.thisTranslation["note_saveAs"],
-                defaultName,
-                "UniqueBible.app Note Files (*.uba);;HTML Files (*.html);;HTM Files (*.htm);;All Files (*)", "", options)
-        if fileName:
-            self.saveAsNote(fileName)
-
-    def saveAsNote(self, fileName):
-        if self.html:
-            note = self.editor.toHtml()
-        else:
-            note = self.editor.toPlainText()
-        note = self.fixNoteFont(note)
-        f = open(fileName, "w", encoding="utf-8")
-        f.write(note)
-        f.close()
-        self.noteFileName = fileName
-        self.parent.addExternalFileHistory(fileName)
-        self.parent.setExternalFileButton()
-        self.parent.noteSaved = True
-        self.updateWindowTitle()
-
-    def fixNoteFont(self, note):
-        return re.sub("<body style={0}[ ]*?font-family:[ ]*?'[^']*?';[ ]*?font-size:[ ]*?[0-9]+?pt;".format('"'), "<body style={0}font-family:'{1}'; font-size:{2}pt;".format('"', config.font, config.fontSize), note)
-
-    # formatting styles
-    def format_clear(self):
-        selectedText = self.editor.textCursor().selectedText()
-        if self.html:
-            selectedText = """<span style="font-family:'{0}'; font-size:{1}pt;">{2}</span>""".format(config.font, config.fontSize, selectedText)
-            self.editor.insertHtml(selectedText)
-        else:
-            selectedText = re.sub("<[^\n<>]*?>", "", selectedText)
-            self.editor.insertPlainText(selectedText)
-
-    def format_header1(self):
-        if self.html:
-            self.editor.insertHtml("<h1>{0}</h1>".format(self.editor.textCursor().selectedText()))
-        else:
-            self.editor.insertPlainText("<h1>{0}</h1>".format(self.editor.textCursor().selectedText()))
-
-    def format_header2(self):
-        if self.html:
-            self.editor.insertHtml("<h2>{0}</h2>".format(self.editor.textCursor().selectedText()))
-        else:
-            self.editor.insertPlainText("<h2>{0}</h2>".format(self.editor.textCursor().selectedText()))
-
-    def format_header3(self):
-        if self.html:
-            self.editor.insertHtml("<h3>{0}</h3>".format(self.editor.textCursor().selectedText()))
-        else:
-            self.editor.insertPlainText("<h3>{0}</h3>".format(self.editor.textCursor().selectedText()))
-
-    def format_font(self):
-        ok, font = QFontDialog.getFont(QFont(config.font, config.fontSize), self)
-        if ok:
-            if self.html:
-                self.editor.setCurrentFont(font)
-            else:
-                fontFamily, fontSize, i1, i2, fontWeight, italic, underline, strikeout, *_ = font.key().split(",")
-                spanTag = """<span style="font-family:'{0}'; font-size:{1}pt;""".format(fontFamily, fontSize)
-                # add font weight
-                if fontWeight == "25":
-                    spanTag += " font-weight:200;"
-                elif fontWeight == "75":
-                    spanTag += " font-weight:600;"
-                # add italic style
-                if italic == "1":
-                    spanTag += " font-style:italic;"
-                # add both underline and strikeout style
-                if underline == "1" and strikeout == "1":
-                    spanTag += " text-decoration: underline line-through;"
-                # add underline style
-                elif underline == "1":
-                    spanTag += " text-decoration: underline;"
-                # add strikeout style
-                elif strikeout == "1":
-                    spanTag += " text-decoration: line-through;"
-                # close tag
-                spanTag += '">'
-                self.editor.insertPlainText("{0}{1}</span>".format(spanTag, self.editor.textCursor().selectedText()))
-
-    def format_textColor(self):
-        color = QColorDialog.getColor(Qt.darkRed, self)
-        if color.isValid():
-            if self.html:
-                self.editor.setTextColor(color)
-            else:
-                self.editor.insertPlainText('<span style="color:{0};">{1}</span>'.format(color.name(), self.editor.textCursor().selectedText()))
-
-    def format_textBackgroundColor(self):
-        color = QColorDialog.getColor(Qt.yellow, self)
-        if color.isValid():
-            if self.html:
-                self.editor.setTextBackgroundColor(color)
-            else:
-                self.editor.insertPlainText('<span style="background-color:{0};">{1}</span>'.format(color.name(), self.editor.textCursor().selectedText()))
-
-    def format_bold(self):
-        if self.html:
-            # Reference: https://doc.qt.io/qt-5/qfont.html#Weight-enum
-            # Bold = 75
-            self.editor.setFontWeight(75)
-        else:
-            self.editor.insertPlainText("<b>{0}</b>".format(self.editor.textCursor().selectedText()))
-
-    def format_italic(self):
-        if self.html:
-            self.editor.setFontItalic(True)
-        else:
-            self.editor.insertPlainText("<i>{0}</i>".format(self.editor.textCursor().selectedText()))
-
-    def format_underline(self):
-        if self.html:
-            self.editor.setFontUnderline(True)
-        else:
-            self.editor.insertPlainText("<u>{0}</u>".format(self.editor.textCursor().selectedText()))
-
-    def format_center(self):
-        if self.html:
-            self.editor.setAlignment(Qt.AlignCenter)
-        else:
-            self.editor.insertPlainText("<div style='text-align:center;'>{0}</div>".format(self.editor.textCursor().selectedText()))
-
-    def format_justify(self):
-        if self.html:
-            self.editor.setAlignment(Qt.AlignJustify)
-        else:
-            self.editor.insertPlainText("<div style='text-align:justify;'>{0}</div>".format(self.editor.textCursor().selectedText()))
-
-    def format_left(self):
-        if self.html:
-            self.editor.setAlignment(Qt.AlignLeft)
-        else:
-            self.editor.insertPlainText("<div style='text-align:left;'>{0}</div>".format(self.editor.textCursor().selectedText()))
-
-    def format_right(self):
-        if self.html:
-            self.editor.setAlignment(Qt.AlignRight)
-        else:
-            self.editor.insertPlainText("<div style='text-align:right;'>{0}</div>".format(self.editor.textCursor().selectedText()))
-
-    def format_custom(self):
-        selectedText = self.editor.textCursor().selectedText()
-        selectedText = self.customFormat(selectedText)
-        if self.html:
-            self.editor.insertHtml(selectedText)
-        else:
-            self.editor.insertPlainText(selectedText)
-
-    def customFormat(self, text):
-        # QTextEdit's line break character by pressing ENTER in plain & html mode " "
-        # please note that " " is not an empty string
-        text = text.replace(" ", "\n")
-
-        text = re.sub("^\*[0-9]+? (.*?)$", r"<ol><li>\1</li></ol>", text, flags=re.M)
-        text = text.replace("</ol>\n<ol>", "\n")
-        text = re.sub("^\* (.*?)$", r"<ul><li>\1</li></ul>", text, flags=re.M)
-        text = text.replace("</ul>\n<ul>", "\n")
-        text = re.sub("^{.*?}$", self.formatHTMLTable, text, flags=re.M)
-        text = text.replace("</table>\n<table>", "\n")
-
-        # add style to table here
-        # please note that QTextEdit supports HTML 4, rather than HTML 5
-        # take this old reference: https://www.w3schools.com/tags/tag_table.asp
-        text = text.replace('<table>', '<table border="1" cellpadding="5">')
-
-        # convert back to QTextEdit linebreak
-        text = text.replace("\n", " ")
-
-        # wrap with default font and font-size
-        text = """<span style="font-family:'{0}'; font-size:{1}pt;">{2}</span>""".format(config.font, config.fontSize, text)
-        return text
-
-    def formatHTMLTable(self, match):
-        row = match.group()[1:-1]
-        row = "".join(["<td>{0}</td>".format(cell) for cell in row.split("|")])
-        return "<table><tr>{0}</tr></table>".format(row)
-
-    def addInternalImage(self):
-        self.openImageDialog(external=False)
-
-    def openImageDialog(self, external=True):
-        options = QFileDialog.Options()
-        fileName, filtr = QFileDialog.getOpenFileName(self,
-                config.thisTranslation["html_open"],
-                self.parent.openFileNameLabel.text(),
-                "JPG Files (*.jpg);;JPEG Files (*.jpeg);;PNG Files (*.png);;GIF Files (*.gif);;BMP Files (*.bmp);;All Files (*)", "", options)
-        if fileName:
-            if external:
-                self.linkExternalImage(fileName)
-            else:
-                self.embedImage(fileName)
-
-    def embedImage(self, fileName):
-        name, extension = os.path.splitext(os.path.basename(fileName))
-        with open(fileName, "rb") as fileObject:
-            binaryData = fileObject.read()
-            encodedData = base64.b64encode(binaryData)
-            asciiString = encodedData.decode('ascii')
-            imageTag = '<img src="data:image/{2};base64,{0}" alt="{1}">'.format(asciiString, name, extension[1:])
-            if self.html:
-                self.editor.insertHtml(imageTag)
-            else:
-                self.editor.insertPlainText(imageTag)
-
-    def exportNoteImages(self):
-        options = QFileDialog.DontResolveSymlinks | QFileDialog.ShowDirsOnly
-        directory = QFileDialog.getExistingDirectory(self,
-                config.thisTranslation["select_a_folder"],
-                self.parent.directoryLabel.text(), options)
-        if directory:
-            if self.html:
-                htmlText = self.editor.toHtml()
-            else:
-                htmlText = self.editor.toPlainText()
-            searchPattern = r'src=(["{0}])data:image/([^<>]+?);[ ]*?base64,[ ]*?([^ <>]+?)\1'.format("'")
-            for counter, value in enumerate(re.findall(searchPattern, htmlText)):
-                *_, ext, asciiString = value
-                binaryString = asciiString.encode("ascii")
-                binaryData = base64.b64decode(binaryString)
-                imageFilePath = os.path.join(directory, "image{0}.{1}".format(counter + 1, ext))
-                with open(imageFilePath, "wb") as fileObject:
-                    fileObject.write(binaryData)
-
-    def linkExternalImage(self, fileName):
-        imageTag = '<img src="{0}" alt="UniqueBible.app">'.format(fileName)
-        if self.html:
-            self.editor.insertHtml(imageTag)
-        else:
-            self.editor.insertPlainText(imageTag)
-
-    def addHyperlink(self, hyperlink):
-        hyperlink = '<a href="{0}">{1}</a>'.format(hyperlink, self.editor.textCursor().selectedText())
-        hyperlink = """<span style="font-family:'{0}'; font-size:{1}pt;">{2}</span>""".format(config.font, config.fontSize, hyperlink)
-        if self.html:
-            self.editor.insertHtml(hyperlink)
-        else:
-            self.editor.insertPlainText(hyperlink)
-
-    def openHyperlinkDialog(self):
-        selectedText = self.editor.textCursor().selectedText()
-        if selectedText:
-            hyperlink = selectedText
-        else:
-            hyperlink = "https://BibleTools.app"
-        text, ok = QInputDialog.getText(self, "UniqueBible.app",
-                config.thisTranslation["noteTool_hyperlink"], QLineEdit.Normal,
-                hyperlink)
-        if ok and text != '':
-            self.addHyperlink(text)
-
-
-class ImportSettings(QDialog):
-
-    def __init__(self, parent):
-        super().__init__()
-        self.parent = parent
-        #self.setModal(True)
-
-        self.setWindowTitle(config.thisTranslation["menu8_settings"])
-
-        self.setupLayout()
-
-    def setupLayout(self):
-        titleBibles = QLabel(config.thisTranslation["menu5_bible"])
-
-        self.linebreak = QCheckBox()
-        self.linebreak.setText(config.thisTranslation["import_linebreak"])
-        self.linebreak.setChecked(config.importAddVerseLinebreak)
-
-        self.stripStrNo = QCheckBox()
-        self.stripStrNo.setText(config.thisTranslation["import_strongNo"])
-        self.stripStrNo.setChecked(config.importDoNotStripStrongNo)
-
-        self.stripMorphCode = QCheckBox()
-        self.stripMorphCode.setText(config.thisTranslation["import_morphCode"])
-        self.stripMorphCode.setChecked(config.importDoNotStripMorphCode)
-
-        self.importRtlOT = QCheckBox()
-        self.importRtlOT.setText(config.thisTranslation["import_rtl"])
-        self.importRtlOT.setChecked(config.importRtlOT)
-
-        self.importInterlinear = QCheckBox()
-        self.importInterlinear.setText(config.thisTranslation["import_interlinear"])
-        self.importInterlinear.setChecked(config.importInterlinear)
-
-        saveButton = QPushButton(config.thisTranslation["note_save"])
-        saveButton.clicked.connect(self.saveSettings)
-
-        cancelButton = QPushButton(config.thisTranslation["message_cancel"])
-        cancelButton.clicked.connect(self.close)
-
-        self.layout = QGridLayout()
-        self.layout.addWidget(titleBibles, 0, 0)
-        self.layout.addWidget(self.linebreak, 1, 0)
-        self.layout.addWidget(self.stripStrNo, 2, 0)
-        self.layout.addWidget(self.stripMorphCode, 3, 0)
-        self.layout.addWidget(self.importRtlOT, 4, 0)
-        self.layout.addWidget(self.importInterlinear, 5, 0)
-        self.layout.addWidget(saveButton, 6, 0)
-        self.layout.addWidget(cancelButton, 7, 0)
-        self.setLayout(self.layout)
-
-    def saveSettings(self):
-        if self.linebreak.isChecked():
-            config.importAddVerseLinebreak = True
-        else:
-            config.importAddVerseLinebreak = False
-
-        if self.stripStrNo.isChecked():
-            config.importDoNotStripStrongNo = True
-        else:
-            config.importDoNotStripStrongNo = False
-
-        if self.stripMorphCode.isChecked():
-            config.importDoNotStripMorphCode = True
-        else:
-            config.importDoNotStripMorphCode = False
-
-        if self.importRtlOT.isChecked():
-            config.importRtlOT = True
-        else:
-            config.importRtlOT = False
-
-        if self.importInterlinear.isChecked():
-            config.importInterlinear = True
-        else:
-            config.importInterlinear = False
-
-        self.close()
-
-
-class MoreConfigOptions(QDialog):
-
-    def __init__(self, parent):
-        super().__init__()
-        self.parent = parent
-        #self.setModal(True)
-        self.setWindowTitle(config.thisTranslation["menu1_moreConfig"])
-        self.setupLayout()
-
-    def setupLayout(self):
-        layout = QVBoxLayout()
-
-        readWiki = QLabel(config.thisTranslation["message_readWiki"])
-        layout.addWidget(readWiki)
-        readWiki = QLabel("https://github.com/eliranwong/UniqueBible/wiki/config_file")
-        layout.addWidget(readWiki)
-
-        horizontalContainer = QWidget()
-        horizontalContainer.setPalette(Themes.getPalette(config.theme))
-        horizontalContainerLayout = QHBoxLayout()
-
-        leftContainer = QWidget()
-        leftContainerLayout = QVBoxLayout()
-        rightContainer = QWidget()
-        rightContainerLayout = QVBoxLayout()
-
-        options = [
-            ("openBibleWindowContentOnNextTab", config.openBibleWindowContentOnNextTab, self.openBibleWindowContentOnNextTabChanged),
-            ("openStudyWindowContentOnNextTab", config.openStudyWindowContentOnNextTab, self.openStudyWindowContentOnNextTabChanged),
-            ("showVerseNumbersInRange", config.showVerseNumbersInRange, self.showVerseNumbersInRangeChanged),
-            ("addFavouriteToMultiRef", config.addFavouriteToMultiRef, self.addFavouriteToMultiRefChanged),
-            ("alwaysDisplayStaticMaps", config.alwaysDisplayStaticMaps, self.alwaysDisplayStaticMapsChanged),
-            ("exportEmbeddedImages", config.exportEmbeddedImages, self.exportEmbeddedImagesChanged),
-            ("showNoteIndicatorOnBibleChapter", config.showNoteIndicatorOnBibleChapter, self.parent.enableNoteIndicatorButtonClicked),
-            ("clickToOpenImage", config.clickToOpenImage, self.clickToOpenImageChanged),
-            ("overwriteNoteFont", config.overwriteNoteFont, self.overwriteNoteFontChanged),
-            ("overwriteBookFont", config.overwriteBookFont, self.overwriteBookFontChanged),
-            ("overwriteNoteFontSize", config.overwriteNoteFontSize, self.overwriteNoteFontSizeChanged),
-            ("overwriteBookFontSize", config.overwriteBookFontSize, self.overwriteBookFontSizeChanged),
-            ("openBibleNoteAfterSave", config.openBibleNoteAfterSave, self.openBibleNoteAfterSaveChanged),
-            ("bookOnNewWindow", config.bookOnNewWindow, self.bookOnNewWindowChanged),
-            ("parserStandarisation", (config.parserStandarisation == "YES"), self.parserStandarisationChanged),
-            ("virtualKeyboard", config.virtualKeyboard, self.virtualKeyboardChanged),
-            ("showGoogleTranslateEnglishOptions", config.showGoogleTranslateEnglishOptions, self.showGoogleTranslateEnglishOptionsChanged),
-            ("autoCopyGoogleTranslateOutput", config.autoCopyGoogleTranslateOutput, self.autoCopyGoogleTranslateOutputChanged),
-            ("showGoogleTranslateChineseOptions", config.showGoogleTranslateChineseOptions, self.showGoogleTranslateChineseOptionsChanged),
-            ("autoCopyChinesePinyinOutput", config.autoCopyChinesePinyinOutput, self.autoCopyChinesePinyinOutputChanged),
-            ("disableModulesUpdateCheck", config.disableModulesUpdateCheck, self.disableModulesUpdateCheckChanged)
-        ]
-        if platform.system() == "Linux":
-            options += [
-                ("linuxStartFullScreen", config.linuxStartFullScreen, self.linuxStartFullScreenChanged),
-                ("showTtsOnLinux", config.showTtsOnLinux, self.showTtsOnLinuxChanged),
-                ("ibus", config.ibus, self.ibusChanged),
-            ]
-        for counter, content in enumerate(options):
-            name, value, function = content
-            checkbox = QCheckBox()
-            checkbox.setText(name)
-            checkbox.setChecked(value)
-            checkbox.stateChanged.connect(function)
-            if counter % 2 == 0:
-                leftContainerLayout.addWidget(checkbox)
-            else:
-                rightContainerLayout.addWidget(checkbox)
-
-        leftContainer.setLayout(leftContainerLayout)
-        rightContainer.setLayout(rightContainerLayout)
-        horizontalContainerLayout.addWidget(leftContainer)
-        horizontalContainerLayout.addWidget(rightContainer)
-        horizontalContainer.setLayout(horizontalContainerLayout)
-        layout.addWidget(horizontalContainer)
-        self.setLayout(layout)
-
-    def virtualKeyboardChanged(self):
-        config.virtualKeyboard = not config.virtualKeyboard
-        self.parent.displayMessage(config.thisTranslation["message_restart"])
-
-    def showVerseNumbersInRangeChanged(self):
-        config.showVerseNumbersInRange = not config.showVerseNumbersInRange
-
-    def openBibleWindowContentOnNextTabChanged(self):
-        config.openBibleWindowContentOnNextTab = not config.openBibleWindowContentOnNextTab
-        self.newTabException = False
-
-    def openStudyWindowContentOnNextTabChanged(self):
-        config.openStudyWindowContentOnNextTab = not config.openStudyWindowContentOnNextTab
-        self.newTabException = False
-
-    def addFavouriteToMultiRefChanged(self):
-        config.addFavouriteToMultiRef = not config.addFavouriteToMultiRef
-
-    def exportEmbeddedImagesChanged(self):
-        config.exportEmbeddedImages = not config.exportEmbeddedImages
-
-    def clickToOpenImageChanged(self):
-        config.clickToOpenImage = not config.clickToOpenImage
-
-    def bookOnNewWindowChanged(self):
-        config.bookOnNewWindow = not config.bookOnNewWindow
-
-    def overwriteNoteFontChanged(self):
-        config.overwriteNoteFont = not config.overwriteNoteFont
-
-    def overwriteNoteFontSizeChanged(self):
-        config.overwriteNoteFontSize = not config.overwriteNoteFontSize
-
-    def overwriteBookFontChanged(self):
-        config.overwriteBookFont = not config.overwriteBookFont
-
-    def overwriteBookFontSizeChanged(self):
-        config.overwriteBookFontSize = not config.overwriteBookFontSize
-
-    def alwaysDisplayStaticMapsChanged(self):
-        config.alwaysDisplayStaticMaps = not config.alwaysDisplayStaticMaps
-
-    def openBibleNoteAfterSaveChanged(self):
-        config.openBibleNoteAfterSave = not config.openBibleNoteAfterSave
-
-    def showGoogleTranslateEnglishOptionsChanged(self):
-        config.showGoogleTranslateEnglishOptions = not config.showGoogleTranslateEnglishOptions
-        self.parent.displayMessage(config.thisTranslation["message_restart"])
-
-    def showGoogleTranslateChineseOptionsChanged(self):
-        config.showGoogleTranslateChineseOptions = not config.showGoogleTranslateChineseOptions
-        self.parent.displayMessage(config.thisTranslation["message_restart"])
-
-    def autoCopyGoogleTranslateOutputChanged(self):
-        config.autoCopyGoogleTranslateOutput = not config.autoCopyGoogleTranslateOutput
-
-    def autoCopyChinesePinyinOutputChanged(self):
-        config.autoCopyChinesePinyinOutput = not config.autoCopyChinesePinyinOutput
-
-    def disableModulesUpdateCheckChanged(self):
-        config.disableModulesUpdateCheck = not config.disableModulesUpdateCheck
-
-    def parserStandarisationChanged(self):
-        if config.parserStandarisation == "YES":
-            config.parserStandarisation = "NO"
-        else:
-            config.parserStandarisation = "YES"
-
-    def linuxStartFullScreenChanged(self):
-        config.linuxStartFullScreen = not config.linuxStartFullScreen
-        self.parent.displayMessage(config.thisTranslation["message_restart"])
-
-    def showTtsOnLinuxChanged(self):
-        config.showTtsOnLinux = not config.showTtsOnLinux
-        self.parent.displayMessage(config.thisTranslation["message_restart"])
-
-    def ibusChanged(self):
-        config.ibus = not config.ibus
-        self.parent.displayMessage(config.thisTranslation["message_restart"])
-
-
-class Downloader(QDialog):
-
-    def __init__(self, parent, databaseInfo):
-        super().__init__()
-        self.parent = parent
-        self.setWindowTitle(config.thisTranslation["message_downloadHelper"])
-        self.setModal(True)
-
-        self.databaseInfo = databaseInfo
-        fileItems, cloudID, *_ = databaseInfo
-        self.cloudFile = "https://drive.google.com/uc?id={0}".format(cloudID)
-        self.localFile = "{0}.zip".format(os.path.join(*fileItems))
-        self.filename = fileItems[-1]
-
-        self.setupLayout()
-
-    def setupLayout(self):
-
-        #self.setupProgressBar()
-
-        message = QLabel("{1} '{0}'".format(self.filename, config.thisTranslation["message_missing"]))
-
-        downloadButton = QPushButton(config.thisTranslation["message_install"])
-        downloadButton.clicked.connect(self.startDownloadFile)
-
-        cancelButton = QPushButton(config.thisTranslation["message_cancel"])
-        cancelButton.clicked.connect(self.close)
-
-        remarks = QLabel("{0} {1}".format(config.thisTranslation["message_remarks"], config.thisTranslation["message_downloadAllFiles"]))
-
-        self.layout = QGridLayout()
-        #self.layout.addWidget(self.progressBar, 0, 0)
-        self.layout.addWidget(message, 0, 0)
-        self.layout.addWidget(downloadButton, 1, 0)
-        self.layout.addWidget(cancelButton, 2, 0)
-        self.layout.addWidget(remarks, 3, 0)
-        self.setLayout(self.layout)
-
-    def setupProgressBar(self):
-        self.progressBar = QProgressBar()
-        self.progressBar.setMinimum(0)
-        self.progressBar.setMaximum(100)
-        self.progressBar.setValue(0)
-
-    def startDownloadFile(self):
-        self.downloadFile(True)
-
-    def downloadFile(self, interactWithParent=True):
-        try:
-            gdown.download(self.cloudFile, self.localFile, quiet=True)
-            connection = True
-        except:
-            #connection = False
-            try:
-                gdown.download(self.cloudFile, self.localFile, quiet=True, proxy=None)
-                connection = True
-            except:
-                connection = False
-        if connection:
-            if self.localFile.endswith(".zip"):
-                zipObject = zipfile.ZipFile(self.localFile, "r")
-                path, *_ = os.path.split(self.localFile)
-                zipObject.extractall(path)
-                zipObject.close()
-                os.remove(self.localFile)
-            if interactWithParent:
-                self.parent.moduleInstalled(self.databaseInfo)
-        else:
-            if interactWithParent:
-                self.parent.moduleInstalledFailed(self.databaseInfo)
