+++ gui/imports.py
@@ -1,0 +1,14 @@
+try:
+    from PySide2.QtTextToSpeech import QTextToSpeech, QVoice
+except:
+    pass
+
+try:
+    from pypinyin import pinyin
+except:
+    pass
+
+try:
+    from googletrans import Translator
+except:
+    pass
